% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{append_table}
\alias{append_table}
\title{Append a Table}
\usage{
append_table(
  conn,
  conn_fun,
  schema,
  tableName,
  data,
  verbose = TRUE,
  render_sql = TRUE,
  ...
)
}
\arguments{
\item{...}{Additional arguments passed to \code{DatabaseConnector::dbAppendTable()}}
}
\description{
Like the writeTable function, this function is a wrapper around a DatabaseConnector function rather than one where a SQL statement is rendered using the SqlRender package. This function performs the additional step of converting all inputs to the data.frame class, especially in cases where the input is a tibble.
}
\seealso{
Other table functions: 
\code{\link{search_table}()}
}
\concept{table functions}
