% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decamel-deprec.R
\name{append_table2}
\alias{append_table2}
\title{Append to am Existing Table}
\usage{
append_table2(conn = conn, schema, tableName, .data, ...)
}
\arguments{
\item{schema}{schema where \code{tableName} is located.}

\item{tableName}{Name of table to write to.}

\item{.data}{dataframe to append}

\item{...}{Additional arguments passed to DatabaseConnector::dbAppendTable}
}
\description{
Like the writeTable function, this function is a wrapper around a DatabaseConnector function rather than one where a SQL statement is rendered using the SqlRender package. This function performs the additional step of converting all inputs to the data.frame class, especially in cases where the input is a tibble.
}
