% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{appendTable}
\alias{appendTable}
\title{Append a Table}
\usage{
appendTable(
  conn,
  conn_fun,
  schema,
  tableName,
  data,
  verbose = TRUE,
  render_sql = TRUE,
  ...
)
}
\arguments{
\item{...}{Additional arguments passed to \code{DatabaseConnector::dbAppendTable()}}
}
\description{
Like the writeTable function, this function is a wrapper around a DatabaseConnector function rather than one where a SQL statement is rendered using the SqlRender package. This function performs the additional step of converting all inputs to the data.frame class, especially in cases where the input is a tibble.
}
\seealso{
Other table functions: 
\code{\link{append_table}()},
\code{\link{create_table}()},
\code{\link{drop_table}()},
\code{\link{read_table}()},
\code{\link{searchTable}()},
\code{\link{search_table}()},
\code{\link{write_table}()}
}
\concept{table functions}
