% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brake.R
\name{brake_closed_conn}
\alias{brake_closed_conn}
\title{Brake if the Connection is Closed}
\usage{
brake_closed_conn(conn)
}
\arguments{
\item{conn}{Connection object}
}
\value{
Halted execution of a function or other process
}
\description{
Stop an operation if the supplied Connection Class object is closed.
}
\details{
This function is derived from the \code{\link{brake_ff}}.
}
\seealso{
\code{\link{brake_ff}}
}
