% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{brake_ff}
\alias{brake_ff}
\title{Create a Brake Function}
\usage{
brake_ff(predicate, stop_message, arguments)
}
\arguments{
\item{predicate}{Evaluating function.}

\item{stop_message}{Message to receive with the \code{predicate} evaluates to true.}

\item{arguments}{(option) character vector of arguments for the new function. If missing, the new function will not have any arguments. Arguments in the \code{predicate} should also be considered. For example, if the predicate is "is.logical(x)", the \code{arguments} value should be "x".}
}
\value{
Assigned function
}
\description{
Author new functions that stops and provides a custom message when a predicate evaluates to true.
}
\seealso{
\code{\link[rlang]{parse_expr}},\code{\link[rlang]{missing_arg}}
\code{\link[purrr]{map}},\code{\link[purrr]{set_names}}

Other ff functions: 
\code{\link{flag_ff}()}
}
\concept{brake functions}
\concept{ff functions}
