% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_join_query}
\alias{build_join_query}
\title{Build a Query for a Join}
\usage{
build_join_query(
  select_table_fields = "*",
  select_join_on_fields = "*",
  distinct = FALSE,
  schema,
  table,
  column,
  join_on_schema,
  join_on_table,
  join_on_column,
  kind = c("LEFT", "RIGHT", "INNER", "FULL"),
  where_in_field,
  where_in_vector,
  where_in_join_on_field,
  where_in_join_on_vector,
  where_not_in_field,
  where_not_in_vector,
  where_not_in_join_on_field,
  where_not_in_join_on_vector,
  where_is_null_field,
  where_is_not_null_field,
  where_is_null_join_on_field,
  where_is_not_null_join_on_field,
  case_insensitive,
  limit,
  random
)
}
\arguments{
\item{select_table_fields}{The fields to select for in the first table}

\item{select_join_on_fields}{The fields to select for in the table being join to the first table}

\item{distinct}{If TRUE, the distinct row count will be returned.}

\item{schema}{The target schema for the operation.}

\item{table}{Target table for the operation.}

\item{column}{Column to join on.}

\item{join_on_schema}{Schema of the table that is being joined to the first table.}

\item{join_on_table}{Table that is being joined to the first table.}

\item{join_on_column}{Column in the \code{join_on_table} that is joined onto the \code{column} of the first table.}

\item{kind}{Type of join. Defaults to left, and options include "LEFT", "RIGHT", "INNER", or "FULL"}

\item{where_in_field}{Paired with \code{where_in_vector}, adds a "WHERE {field} IN ({vector})" clause to the query corresponding to the field-value pair.}

\item{where_not_in_field}{Paired with \code{where_not_in_vector}, adds a "WHERE {field} NOT IN ({vector})" clause to the query corresponding to the field-value pair.}

\item{case_insensitive}{If TRUE, both sides of the query are converted to lowercase.}

\item{limit}{(Optional) Integer of the row limit. Takes precedence over \code{random} if both arguments are provided.}

\item{random}{(Optional) Integer of the random number of rows to return. Is preceded by \code{limit} if both \code{limit} and \code{random} are provided.}
}
\description{
Build a query the joins 2 tables, respectively aliased as \code{a} and \code{b} in the query itself.
}
\seealso{
\code{\link[cli]{cli_alert}}
\code{\link[SqlRender]{render}}
}
