% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_query_like}
\alias{build_query_like}
\title{Build a Like Query}
\usage{
build_query_like(
  fields = "*",
  distinct = FALSE,
  schema,
  table,
  where_like_field,
  where_like_value,
  case_insensitive = TRUE,
  limit_n = NULL
)
}
\arguments{
\item{fields}{Fields selected for. Defaults to "*".}

\item{distinct}{If TRUE, the distinct row count will be returned.}

\item{schema}{The target schema for the operation.}

\item{table}{Target table for the operation.}

\item{case_insensitive}{If TRUE, both sides of the query are converted to lowercase.}
}
\description{
A \verb{Like Query} is one where a search for a value is performed with flanking wildcards.
}
\seealso{
\code{\link[SqlRender]{render}}
}
