% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_query_string}
\alias{build_query_string}
\title{Build a Query the Loops Over String}
\usage{
build_query_string(
  fields = "*",
  distinct = FALSE,
  schema,
  table,
  where_like_field,
  string,
  split = " |[[:punct:]]",
  case_insensitive = TRUE,
  limit_n = NULL
)
}
\description{
Write a SQL Query that loops over the words in a string.
}
\seealso{
\code{\link[SqlRender]{render}}
}
