% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildJoinQuery.R
\name{buildJoinQuery}
\alias{buildJoinQuery}
\title{Build a SQL Query}
\usage{
buildJoinQuery(
  fields = "*",
  distinct = FALSE,
  schema,
  tableName,
  column,
  joinSchema,
  joinTableName,
  joinColumn,
  joinType,
  whereInField = NULL,
  whereInVector = NULL,
  whereNotInField = NULL,
  whereNotInVector = NULL,
  caseInsensitive = TRUE,
  n = NULL,
  n_type = c("limit", "random")
)
}
\value{
SQL statement as a character string.
}
\description{
A SQL query is built using the given arguments. Currently, only 1 whereIn and whereNot in parameters can be set.
}
