% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildQuery.R
\name{buildQuery}
\alias{buildQuery}
\title{Build a Query based on Constructs}
\usage{
buildQuery(
  fields = "*",
  schema,
  tableName,
  whereInField = NULL,
  whereInVector = NULL,
  whereNotInField = NULL,
  whereNotInVector = NULL,
  caseInsensitive = TRUE,
  n = NULL,
  n_type = c("limit", "random")
)
}
\description{
A SQL query is built from construction of parts from the construct primitive functions in this package.
}
