% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{buildQueryLike}
\alias{buildQueryLike}
\title{Writes a Like SQL Query}
\usage{
buildQueryLike(
  fields = "*",
  distinct = FALSE,
  schema,
  tableName,
  whereLikeField,
  whereLikeValue,
  caseInsensitive = TRUE,
  limit_n = NULL
)
}
\value{
a SQL Query as a character string.
}
\description{
Writes a Like SQL Query
}
\keyword{internal}
