% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildQueryString.R
\name{buildQueryString}
\alias{buildQueryString}
\title{Write a SQL statement that loops over a word in a string}
\usage{
buildQueryString(
  fields = "*",
  schema,
  tableName,
  whereLikeField,
  string,
  split,
  caseInsensitive = TRUE,
  limit_n = NULL
)
}
\description{
Write a SQL statement that loops over a word in a string
}
