% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R, R/decamel-cache.R,
%   R/decamel-decamel-cache.R
\name{cache}
\alias{cache}
\title{Cache a Resultset}
\usage{
cache(object, sql_statement, dirs)

cache(object, sql_statement, dirs)

cache(object, sql_statement, dirs)
}
\arguments{
\item{object}{R object to cache.}

\item{sql_statement}{SQL Statement from which a hexadecimal hash code will be generated and appended to the cache filename.}

\item{dirs}{Argument passed to \code{\link[R.cache]{saveCache}}.}
}
\value{
Invisibly returns the path to the cache file.

Invisibly returns the path to the cache file.

Invisibly returns the path to the cache file.
}
\description{
Cache a resultset with the SQL Statement as the key to a "fantasia" cache directory using a wrapper around the \code{\link[R.cache]{saveCache}} function.

Cache a resultset with the SQL Statement as the key to a "fantasia" cache directory using a wrapper around the \code{\link[R.cache]{saveCache}} function.

Cache a resultset with the SQL Statement as the key to a "fantasia" cache directory using a wrapper around the \code{\link[R.cache]{saveCache}} function.
}
\seealso{
Other cache functions: 
\code{\link{loadCache}()},
\code{\link{load_cache}()}

Other cache functions: 
\code{\link{loadCache}()},
\code{\link{load_cache}()}

Other cache functions: 
\code{\link{loadCache}()},
\code{\link{load_cache}()}
}
\concept{cache functions}
