% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{cache_join}
\alias{cache_join}
\title{Cache a Resultset from a Join}
\usage{
cache_join(.data, ..., db, schema)
}
\arguments{
\item{.data}{dataframe to cache}

\item{...}{Arguments passed to the buildJoinQuery function}

\item{db}{name of database}

\item{schema}{name of schema}
}
\description{
This function will cache the resultset from a SQL query created by buildJoinQuery to a subdirectory of the cache root directory, the path of which will be {db/schema}.
}
