% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{cacheQuery}
\alias{cacheQuery}
\title{Cache a Query Resultset}
\usage{
cacheQuery(.data, sqlQuery, db)
}
\arguments{
\item{.data}{dataframe to cache}

\item{sqlQuery}{sql statement of length 1 that will be the sole key for cache retrieval.}

\item{db}{name of database}

\item{schema}{name of schema}
}
\description{
This function will cache the resultset from a SQL query to a subdirectory of the cache root directory, the path of which will be {db/schema}.
}
