% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_plr_example.R
\name{cat_plr_example}
\alias{cat_plr_example}
\title{Print an Example of a Postgres Function using PLR}
\usage{
cat_plr_example()
}
\description{
Print an Example of a Postgres Function using PLR
}
\seealso{
Other PLR: 
\code{\link{is_plr_installed}()}
}
\concept{PLR}
