% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-typewrite.R
\name{cat_sql}
\alias{cat_sql}
\title{Concatenate SQL statement}
\usage{
cat_sql(
  sql_statement,
  tab_count = 0,
  sep = "\\n",
  sql_file = "",
  append = FALSE
)
}
\arguments{
\item{sql_statement}{SQL statement.}

\item{tab_count}{Number of tabs the chunk will be indented when printed in
the console, Default: 0.}

\item{sep}{Passed to \code{cat()}.}

\item{sql_file}{Passed as \code{file} argument to \verb{cat().} An error will be
thrown if the extension is not \emph{sql}. If value is '""', then
no file is written.}

\item{append}{Passed to \code{cat()}.}
}
\description{
Return a SQL statement to the console. Output will be written
verbatim to an sql file path if provided.
}
