% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-typewrite.R
\name{cat_sql_chunk}
\alias{cat_sql_chunk}
\title{Cat SQL Chunk}
\usage{
cat_sql_chunk(sql_statement, tab_count = 0)
}
\arguments{
\item{sql_statement}{SQL statement.}

\item{tab_count}{Number of tabs the chunk will be indented when printed in
the console, Default: 0.}
}
\description{
Return a given SQL around ticks for R Markdown.
}
