% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_conn}
\alias{check_conn}
\title{Check a Connection Object}
\usage{
check_conn(conn)
}
\description{
Check a Connection Object
}
\seealso{
Other check functions: 
\code{\link{check_inflow}()},
\code{\link{check_outflow}()}
}
\concept{check connection functions}
\concept{check functions}
