% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-checks.R
\name{check_conn_type}
\alias{check_conn_type}
\title{Check if a Connection is JDBC}
\usage{
check_conn_type(conn, log_file = "", sep = "\\n", append = TRUE)
}
\description{
Check if a Connection is JDBC
}
\keyword{internal}
