% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-checks.R
\name{check_field_name}
\alias{check_field_name}
\title{Check a field name}
\usage{
check_field_name(field_name, log_file = "", sep = "\\n", append = TRUE)
}
\description{
Check that a field name is not a reserved SQL word.
}
\keyword{internal}
