% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-checks.R
\name{check_field_names}
\alias{check_field_names}
\title{Check a multiple field names}
\usage{
check_field_names(field_names)
}
\description{
Check that more than 1 field name is not a reserved SQL word.
}
\keyword{internal}
