% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_inflow}
\alias{check_inflow}
\title{Check Incoming Data}
\usage{
check_inflow(data)
}
\description{
Check Incoming Data
}
\seealso{
Other check functions: 
\code{\link{check_conn}()},
\code{\link{check_outflow}()}

Other check data functions: 
\code{\link{check_outflow}()}
}
\concept{check data functions}
\concept{check functions}
