% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_outflow}
\alias{check_outflow}
\title{Check Outgoing Data}
\usage{
check_outflow(data, table_name)
}
\description{
Check Outgoing Data
}
\seealso{
Other check functions: 
\code{\link{check_conn}()},
\code{\link{check_inflow}()}

Other check data functions: 
\code{\link{check_inflow}()}
}
\concept{check data functions}
\concept{check functions}
