% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-checks.R
\name{check_rows}
\alias{check_rows}
\title{Check that the incoming data has rows}
\usage{
check_rows(data, log_file = "", sep = "\\n", append = TRUE)
}
\description{
Check that the incoming data has rows
}
\keyword{internal}
