% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-checks.R
\name{check_source_rows}
\alias{check_source_rows}
\title{Check that the table has any rows}
\usage{
check_source_rows(
  sql_statement,
  conn,
  log_file = "",
  sep = "\\n",
  append = TRUE
)
}
\description{
Check that the table has any rows
}
