% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-checks.R
\name{check_total_rows}
\alias{check_total_rows}
\title{Check that the table has more than 1 row}
\usage{
check_total_rows(
  sql_statement,
  conn,
  log_file = "",
  sep = "\\n",
  append = TRUE
)
}
\description{
Check that the table has more than 1 row
}
\keyword{internal}
