% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decamel-cache_funs.R,
%   R/decamel-decamel-cache_funs.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear a Cache}
\usage{
clear_cache(db)

clear_cache(db)
}
\arguments{
\item{db}{name of database}

\item{.data}{dataframe to cache}

\item{sqlQuery}{sql statement of length 1 that will be the sole key for cache retrieval.}

\item{schema}{name of schema}
}
\description{
This function will clear a cache subdirectory of the cache root directory, the path of which will be {db/schema}.

This function will clear a cache subdirectory of the cache root directory, the path of which will be {db/schema}.
}
