% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn-env.R
\name{close_conn}
\alias{close_conn}
\title{FUNCTION_TITLE}
\usage{
close_conn(..., all = FALSE, verbose = TRUE)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}

\item{all}{Close all connections in the \code{pg13_connection_env} environment?}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rlang]{list2}}
}
