% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{concat_where_constructs}
\alias{concat_where_constructs}
\title{Concatenate 2 WHERE constructs}
\usage{
concat_where_constructs(sql_construct, where_sql_construct_2, ..., AND = TRUE)
}
\arguments{
\item{AND}{If TRUE, the WHERE constructs are concatenated with "AND". If FALSE, the concatenation is performed with an "OR".}
}
\description{
When 2 WHERE constructs are included in the SQL Statement, they are concatenated and the 2nd "WHERE" is replaced with an "AND".
}
