% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatWhereConstructs.R
\name{concatWhereConstructs}
\alias{concatWhereConstructs}
\title{Concatenate multuiple WHERE constructs}
\usage{
concatWhereConstructs(sql_construct, where_sql_construct_2)
}
\description{
When 2 WHERE constructs are included in the SQL Statement, they are concatenated with an "AND"
}
