% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{conn_db}
\alias{conn_db}
\title{Connect to a Postgres Database}
\usage{
conn_db(user, password, port, server)
}
\description{
Connect to a Postgres Database
}
\seealso{
Other connection functions: 
\code{\link{connect_ff}()},
\code{\link{connect}()},
\code{\link{dc0}()},
\code{\link{dc}()},
\code{\link{quietly.conn_db}()}
}
\concept{connection functions}
\keyword{internal}
