% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{connect}
\alias{connect}
\title{Connect to a Postgres Database}
\usage{
connect(
  user,
  password,
  port,
  server,
  extraSettings = NULL,
  oracleDriver = "thin",
  connectionString = NULL,
  pathToDriver = system.file(package = "pg13", "driver"),
  verbose = TRUE
)
}
\arguments{
\item{user}{The user name used to access the server.}

\item{password}{The password for that user.}

\item{port}{(optional) The port on the server to connect to.}

\item{server}{The name of the server.}

\item{extraSettings}{(optional) Additional configuration settings specific to the database
provider to configure things as security for SSL. These must follow the
format for the JDBC connection for the RDBMS specified in dbms.}

\item{oracleDriver}{Specify which Oracle drive you want to use. Choose between \code{"thin"}
or \code{"oci"}.}

\item{connectionString}{The JDBC connection string. If specified, the \code{server}, \code{port},
\code{extraSettings}, and \code{oracleDriver} fields are ignored. If
\code{user} and \code{password} are not specified, they are assumed to
already be included in the connection string.}

\item{pathToDriver}{Path to a folder containing the JDBC driver JAR files. See \code{\link[DatabaseConnector]{downloadJdbcDrivers}} 
for instructions on how to download the relevant drivers.}

\item{verbose}{If TRUE, returns console messages when a connection has been secured.}
}
\value{
An object that extends \code{DBIConnection} in a database-specific manner. This object is used to
direct commands to the database engine.
}
\description{
Connect to a Postgres Database
}
\seealso{
Other connection functions: 
\code{\link{conn_db}()},
\code{\link{connect_ff}()},
\code{\link{dc0}()},
\code{\link{dc}()},
\code{\link{quietly.conn_db}()}
}
\concept{connection functions}
