% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R, R/decamel-connection.R
\name{connect_ff}
\alias{connect_ff}
\title{Connection Function Factory}
\usage{
connect_ff(user, password, port, server)

connect_ff(user, password, port, server)
}
\arguments{
\item{user}{The user name used to access the server.}

\item{password}{The password for that user.}

\item{port}{(optional) The port on the server to connect to.}

\item{server}{The name of the server.}
}
\description{
Customize a connection function.

Customize a connection function.
}
