% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_local_postgres.R
\name{connect_to_local_postgres}
\alias{connect_to_local_postgres}
\title{Connect to Local PostGres Server}
\usage{
connect_to_local_postgres(dbname, schema = NULL, port = "5432")
}
\arguments{
\item{dbname}{database name with the corresponding 'localhost/{dbname}' server}

\item{schema}{defaults to NULL}

\item{port}{defults to "5432"}
}
\description{
Connect to Local PostGres Server
}
