% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_mysql5.5.R
\name{connect_to_mysql5.5}
\alias{connect_to_mysql5.5}
\title{Connect to MySQL v5.5 via Unix Sock using credentials stored in a local .Renviron}
\usage{
connect_to_mysql5.5(dbname)
}
\arguments{
\item{dbname}{database name as character string}
}
\description{
Connect to MySQL v5.5 via Unix Sock using credentials stored in a local .Renviron
}
