% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R, R/decamel-connection.R
\name{connection}
\alias{connection}
\title{Postgres Connection}
\description{
Creating a connection to a Postgres Database involves first making a connectionDetails object, which is a list of credentials and other parameters for the target database, followed by making the database connection itself that is declared in a connection class object.

Creating a connection to a Postgres Database involves first making a connectionDetails object, which is a list of credentials and other parameters for the target database, followed by making the database connection itself that is declared in a connection class object.
}
\seealso{
\code{\link[DatabaseConnector]{createConnectionDetails}},
\code{\link[DatabaseConnector]{connect}}

\code{\link[DatabaseConnector]{createConnectionDetails}},
\code{\link[DatabaseConnector]{connect}}
}
