% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{construct_random}
\alias{construct_random}
\title{Construct ORDER BY RANDOM()}
\usage{
construct_random(n)
}
\arguments{
\item{n}{Row number desired in the output}
}
\description{
Construct ORDER BY RANDOM()
}
\keyword{internal}
