% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decamel-deprec.R
\name{construct_schema_table_name}
\alias{construct_schema_table_name}
\title{Construct schemaTableName}
\usage{
construct_schema_table_name(schema, tableName)
}
\description{
(Deprecated)  construct schemaTableName from the schema and tableName
}
