% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{construct_where_like}
\alias{construct_where_like}
\title{Render "WHERE x LIKE y"}
\usage{
construct_where_like(field, term)
}
\arguments{
\item{field}{Single field to be filtered for}

\item{term}{Character string that the field parameter is searched for.}
}
\description{
(Deprecated)  This is a non-terminal render, meaning that the SQL component will not be terminated with a semicolon in order to construct complex SQL queries.
}
