% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{construct_where_lower_like}
\alias{construct_where_lower_like}
\title{Render WHERE lowercase x LIKE y}
\usage{
construct_where_lower_like(field, term)
}
\arguments{
\item{field}{Single field to be filtered for}

\item{term}{Character string that the field parameter is searched for.}
}
\description{
Render WHERE lowercase x LIKE y
}
\keyword{internal}
