% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructBase.R
\name{constructBase}
\alias{constructBase}
\title{Construct a base SQL query}
\usage{
constructBase(fields = "*", schema = "public", tableName)
}
\description{
Construct the SELECT {fields} FROM {schema}.{table} base query. The SQL Statement is not terminated for the addition of other parameters generated using "construct" functions.
}
