% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec.R
\name{constructJoin}
\alias{constructJoin}
\title{Construct a join part of a SQL Statement}
\usage{
constructJoin(
  schema,
  tableName,
  column,
  joinType = "LEFT",
  joinOnSchema,
  joinOnTableName,
  joinOnColumn
)
}
\description{
(Deprecated)  Construct {join type} JOIN {schema.table.column2} ON {schema.table.column2} = {schema.table.column1} part of the sql statement
}
