% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec.R
\name{constructRandom}
\alias{constructRandom}
\title{Construct ORDER BY RANDOM()}
\usage{
constructRandom(n)
}
\arguments{
\item{n}{Row number desired in the output}
}
\description{
Construct ORDER BY RANDOM()
}
