% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructWhereIn.R
\name{constructWhereIn}
\alias{constructWhereIn}
\title{Render x in "WHERE x IN (y)"}
\usage{
constructWhereIn(field, vector)
}
\arguments{
\item{field}{Single field to be filtered for}

\item{vector}{vector of values that the SQL query is filtering for}
}
\description{
This is a non-terminal render, meaning that the SQL component will not be terminated with a semicolon in order to construct complex SQL queries.
}
