% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decamel-db.R
\name{create_db}
\alias{create_db}
\title{Create a New Database}
\usage{
create_db(conn, dbname, ...)
}
\arguments{
\item{conn}{A connection object returned by \code{connect()}.}

\item{dbname}{Database name.}

\item{...}{Additional arguments passed to \code{DatabaseConnector::dbSendStatement()}.}
}
\description{
Create a New Database
}
