% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{create_log}
\alias{create_log}
\title{Create a Log Table}
\usage{
create_log(
  datetime_field = "log_datetime",
  conn,
  conn_fun,
  schema,
  log_table,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE
)
}
\arguments{
\item{datetime_field}{Name of the datetime field. Defaults to \code{log_datetime}.}

\item{conn}{Connection object}

\item{schema}{The target schema for the operation.}

\item{log_table}{Log table name}

\item{verbose}{If TRUE, details on the activity are returned in the console, such as when the querying starting and was completed.}

\item{render_sql}{If TRUE, the SQL statement for the operation is returned in the console.}
}
\description{
Create a \verb{Log Table} if it does not already exist. A log table consists of a user-defined timestamp field, followed by \code{activity} and \code{status} fields.
}
\seealso{
\code{\link[SqlRender]{render}}
}
