% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_schema.R
\name{create_schema}
\alias{create_schema}
\title{Create schema for a database}
\usage{
create_schema(dbname, new_schema)
}
\arguments{
\item{dbname}{name of postgres database}

\item{new_schema}{name of new schema}
}
\description{
This function connects, creates a new schema after dropping one of it exists, and disconnects.
}
