% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_table}
\alias{create_table}
\title{Create a Table}
\usage{
create_table(
  conn,
  schema,
  table_name,
  if_not_exists = TRUE,
  ...,
  verbose = TRUE,
  render_sql = TRUE
)
}
\arguments{
\item{...}{Named vector of field names and their corresponding data definition.}
}
\description{
Create a Table
}
\seealso{
\code{\link[rlang]{list2}}
\code{\link[SqlRender]{render}}
\code{\link[pg13]{send}}

Other create functions: 
\code{\link{create_table_from_df}()},
\code{\link{draft_create_table_from_df}()},
\code{\link{draft_create_table}()}

Other table functions: 
\code{\link{appendTable}()},
\code{\link{append_table}()},
\code{\link{drop_all_staging_tables}()},
\code{\link{drop_table}()},
\code{\link{read_table}()},
\code{\link{searchTable}()},
\code{\link{search_table}()},
\code{\link{write_staging_table}()},
\code{\link{write_table}()}
}
\concept{create functions}
\concept{table functions}
