% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDB.R
\name{createDB}
\alias{createDB}
\title{Create a new Database}
\usage{
createDB(conn, newDB, ...)
}
\arguments{
\item{...}{Additional arguments passed to the DatabaseConnector::dbSendStatement function}
}
\description{
Create a new Database
}
