% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{draftCreateTable}
\alias{draftCreateTable}
\alias{createTable}
\title{Draft SQL to Create a Table}
\usage{
draftCreateTable(schema, tableName, if_not_exists = TRUE, ...)

createTable(
  conn,
  schema,
  tableName,
  if_not_exists = TRUE,
  ...,
  verbose = TRUE,
  render_sql = TRUE
)
}
\arguments{
\item{...}{Named vector of field names and their corresponding data definition.}
}
\description{
Draft a SQL that creates a table using DDL derived from a dataframe. Drafting is stopped if the given tablename or fields in the dataframe are a SQL reserve word.
}
\seealso{
\code{\link[rlang]{list2}}
\code{\link[SqlRender]{render}}

\code{\link[rlang]{list2}}
\code{\link[SqlRender]{render}}
\code{\link[pg13]{send}}
}
\keyword{internal}
