% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-logical.R
\name{db_exists}
\alias{db_exists}
\title{Does a database exist?}
\usage{
db_exists(conn, conn_fun, db_name)
}
\arguments{
\item{conn}{Connection object}
}
\description{
Does a database exist?
}
\seealso{
Other logical functions: 
\code{\link{field_exists}()},
\code{\link{schema_exists}()},
\code{\link{table_exists}()}
}
\concept{logical functions}
