% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{dc_ff}
\alias{dc_ff}
\title{Write a function that already points to the connection object}
\usage{
dc_ff(conn)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[DatabaseConnector]{character(0)}}
\code{\link[secretary]{typewrite}}
}
