% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_local_conn.R
\name{dc_local_conn}
\alias{dc_local_conn}
\title{Disconnect a connection to a local database}
\usage{
dc_local_conn(conn_to_dc)
}
\description{
This function also removes the connection object from the global environment.
}
