% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disconnect_local_postgres.R
\name{disconnect_local_postgres}
\alias{disconnect_local_postgres}
\title{Disconnect from Postgres}
\usage{
disconnect_local_postgres(conn)
}
\arguments{
\item{conn}{connection object}
}
\description{
Disconnect from Postgres
}
