% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{draft_base}
\alias{draft_base}
\title{Construct a base SQL query}
\usage{
draft_base(fields = "*", distinct = FALSE, schema, table_name)
}
\arguments{
\item{distinct}{if TRUE, the DISTINCT fields will be selected for.}
}
\description{
Construct the SELECT {fields} FROM {schema}.{table} base query. The SQL Statement is not terminated for the addition of other parameters generated using "construct" functions.
}
\seealso{
Other draft functions: 
\code{\link{draft_create_table_from_df}()},
\code{\link{draft_create_table}()},
\code{\link{draft_in}()},
\code{\link{draft_join}()},
\code{\link{draft_limit}()},
\code{\link{draft_random}()},
\code{\link{draft_table_path}()},
\code{\link{draft_where_in}()},
\code{\link{draft_where_like}()},
\code{\link{draft_where_lower_in}()},
\code{\link{draft_where_lower_like}()},
\code{\link{draft_where_lower_not_in}()},
\code{\link{draft_where_not_in}()},
\code{\link{paste_wheres}()},
\code{\link{table.path}()}
}
\concept{draft functions}
