% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decamel-draft.R
\name{draft_join}
\alias{draft_join}
\title{Construct a join part of a SQL Statement}
\usage{
draft_join(
  schema,
  tableName,
  column,
  joinType = "LEFT",
  joinOnSchema,
  joinOnTableName,
  joinOnColumn
)
}
\description{
Construct {join type} JOIN {schema.table.column2} ON {schema.table.column2} = {schema.table.column1} part of the sql statement
}
