% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{draft_random}
\alias{draft_random}
\title{Construct ORDER BY RANDOM()}
\usage{
draft_random(n)
}
\arguments{
\item{n}{Row number desired in the output}
}
\description{
Construct ORDER BY RANDOM()
}
\seealso{
Other draft functions: 
\code{\link{draft_base}()},
\code{\link{draft_create_table_from_df}()},
\code{\link{draft_create_table}()},
\code{\link{draft_in}()},
\code{\link{draft_join}()},
\code{\link{draft_limit}()},
\code{\link{draft_table_path}()},
\code{\link{draft_where_in}()},
\code{\link{draft_where_like}()},
\code{\link{draft_where_lower_in}()},
\code{\link{draft_where_lower_like}()},
\code{\link{draft_where_lower_not_in}()},
\code{\link{draft_where_not_in}()},
\code{\link{paste_wheres}()},
\code{\link{table.path}()}
}
\concept{draft functions}
