% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{draft_table_path}
\alias{draft_table_path}
\title{Construct schemaTableName}
\usage{
draft_table_path(schema, tableName)
}
\description{
construct schemaTableName from the schema and tableName
}
\seealso{
Other draft functions: 
\code{\link{draft_base}()},
\code{\link{draft_create_table_from_df}()},
\code{\link{draft_create_table}()},
\code{\link{draft_in}()},
\code{\link{draft_join}()},
\code{\link{draft_limit}()},
\code{\link{draft_random}()},
\code{\link{draft_where_in}()},
\code{\link{draft_where_like}()},
\code{\link{draft_where_lower_in}()},
\code{\link{draft_where_lower_like}()},
\code{\link{draft_where_lower_not_in}()},
\code{\link{draft_where_not_in}()},
\code{\link{paste_wheres}()},
\code{\link{table.path}()}
}
\concept{draft functions}
