% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{draft_where_like}
\alias{draft_where_like}
\title{Render "WHERE x LIKE y"}
\usage{
draft_where_like(field, term)
}
\arguments{
\item{field}{Single field to be filtered for}

\item{term}{Character string that the field parameter is searched for.}
}
\description{
This is a non-terminal render, meaning that the SQL component will not be terminated with a semicolon in order to construct complex SQL queries.
}
\seealso{
Other draft functions: 
\code{\link{draft_base}()},
\code{\link{draft_create_table_from_df}()},
\code{\link{draft_create_table}()},
\code{\link{draft_in}()},
\code{\link{draft_join}()},
\code{\link{draft_limit}()},
\code{\link{draft_random}()},
\code{\link{draft_table_path}()},
\code{\link{draft_where_in}()},
\code{\link{draft_where_lower_in}()},
\code{\link{draft_where_lower_like}()},
\code{\link{draft_where_lower_not_in}()},
\code{\link{draft_where_not_in}()},
\code{\link{paste_wheres}()},
\code{\link{table.path}()}
}
\concept{draft functions}
