% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{draftBase}
\alias{draftBase}
\title{Construct a base SQL query}
\usage{
draftBase(fields = "*", distinct = FALSE, schema, tableName)
}
\arguments{
\item{distinct}{if TRUE, the DISTINCT fields will be selected for.}
}
\description{
Construct the SELECT {fields} FROM {schema}.{table} base query. The SQL Statement is not terminated for the addition of other parameters generated using "construct" functions.
}
\keyword{internal}
