% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTable.R
\name{draftCreateTableFromDF}
\alias{draftCreateTableFromDF}
\title{FUNCTION_TITLE}
\usage{
draftCreateTableFromDF(schema, tableName, data, if_not_exists = TRUE)
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
\code{\link[forcats]{fct_collapse}}
}
