% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{draftLimit}
\alias{draftLimit}
\title{Construct LIMIT}
\usage{
draftLimit(n)
}
\arguments{
\item{n}{rows to limit to}
}
\description{
This is a non-terminal render, meaning that the SQL component will not be terminated with a semicolon in order to construct complex SQL queries.
}
