% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{draftTablePath}
\alias{draftTablePath}
\title{Construct schemaTableName}
\usage{
draftTablePath(schema, tableName)
}
\description{
construct schemaTableName from the schema and tableName
}
\keyword{internal}
