% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{draftWhereLike}
\alias{draftWhereLike}
\title{Render "WHERE x LIKE y"}
\usage{
draftWhereLike(field, term)
}
\arguments{
\item{field}{Single field to be filtered for}

\item{term}{Character string that the field parameter is searched for.}
}
\description{
This is a non-terminal render, meaning that the SQL component will not be terminated with a semicolon in order to construct complex SQL queries.
}
\keyword{internal}
