% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{draftWhereLowerNotIn}
\alias{draftWhereLowerNotIn}
\title{Render "WHERE lowercase x NOT IN y"}
\usage{
draftWhereLowerNotIn(field, vector)
}
\arguments{
\item{field}{Single field to be filtered for}

\item{vector}{vector of values that the SQL query is filtering for}
}
\description{
This is a non-terminal render, meaning that the SQL component will not be terminated with a semicolon in order to construct complex SQL queries.
}
