% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_funs.R
\name{drop_if_exists}
\alias{drop_if_exists}
\title{Drop a Postgres schema}
\usage{
drop_if_exists(
  conn,
  conn_fun,
  schema,
  cascade = FALSE,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  ...
)
}
\arguments{
\item{...}{Additional arguments passed to the DatabaseConnector::dbSendStatement function}
}
\description{
Drop a schema if it exists.
}
