% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_schema.R
\name{drop_schema}
\alias{drop_schema}
\title{Drop schema for a database}
\usage{
drop_schema(dbname, schema)
}
\arguments{
\item{dbname}{name of postgres database}

\item{schema}{name of schema to drop}
}
\description{
Drop schema for a database
}
