% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_funs.R
\name{dropCascade}
\alias{dropCascade}
\title{Drop a Postgres schema}
\usage{
dropCascade(conn, schema, ...)

dropCascade(conn, schema, ...)
}
\arguments{
\item{...}{Additional arguments passed to the DatabaseConnector::dbSendStatement function}
}
\description{
Drop a schema if it exists.

Drop a schema if it exists.
}
