% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{dropIfExists}
\alias{dropIfExists}
\title{Drop a Postgres schema}
\usage{
dropIfExists(conn, schema, ...)
}
\arguments{
\item{...}{Additional arguments passed to the DatabaseConnector::dbSendStatement function}
}
\description{
Drop a schema if it exists.
}
