% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropTable.R
\name{dropTable}
\alias{dropTable}
\title{Drop a table in a Postgres schema}
\usage{
dropTable(conn, schema, tableName, ...)
}
\arguments{
\item{...}{Additional arguments passed to the DatabaseConnector::dbSendStatement function}
}
\description{
Drop a table if it exists.
}
