% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{execute}
\alias{execute}
\title{Execute SQL}
\usage{
execute(
  conn,
  sql_statement,
  profile = FALSE,
  progressBar = TRUE,
  reportOverallTime = TRUE,
  ...
)
}
\description{
This function differs from the send() and query() functions in that it provides additional features such as a progress bar and time estimations.
}
