% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_statement_to_file.R
\name{export_statement_to_file}
\alias{export_statement_to_file}
\title{Export a sql statement within a R object to a sql file}
\usage{
export_statement_to_file(sql_statement, path_to_sql_file)
}
\arguments{
\item{sql_statement}{sql statement R object}

\item{path_to_sql_file}{sql file path to output. A sql extension is added if it is not present in the input.}
}
\description{
Export a sql statement within a R object to a sql file
}
