% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-logical.R
\name{field_exists}
\alias{field_exists}
\title{Does a field exist?}
\usage{
field_exists(conn, schema, table, field)
}
\arguments{
\item{conn}{Connection object}

\item{schema}{The target schema for the operation.}

\item{table}{Target table for the operation.}

\item{field}{Character string to check for in the given table.}
}
\description{
Logical that checks if a field exists in a table. The \code{field} argument is formatted into lowercase prior to being checked.
}
\seealso{
Other logical functions: 
\code{\link{db_exists}()},
\code{\link{schema_exists}()},
\code{\link{table_exists}()}
}
\concept{logical functions}
