% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{file_send}
\alias{file_send}
\title{Send Query to any Postgres connection}
\usage{
file_send(conn, file, ...)
}
\arguments{
\item{conn}{Connection object}

\item{file}{path to sql file}

\item{...}{Additional arguments to pass onto the DatabaseConnector::dbGetQuery function}

\item{sql_statement}{SQL Query}
}
\description{
Send Query to any Postgres connection
}
