% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{flag_ff}
\alias{flag_ff}
\title{Create a Flag Function}
\usage{
flag_ff(predicate, warn_message, arguments, ...)
}
\arguments{
\item{predicate}{Evaluating function.}

\item{warn_message}{Message to receive with the \code{predicate} evaluates to true.}

\item{arguments}{(option) character vector of arguments for the new function. If missing, the new function will not have any arguments. Arguments in the \code{predicate} should also be considered. For example, if the predicate is "is.logical(x)", the \code{arguments} value should be "x".}

\item{...}{Additional arguments passed to the base warning function}
}
\value{
Assigned function
}
\description{
Author new functions that warns and provides a custom message when a predicate evaluates to true.
}
\seealso{
\code{\link[rlang]{parse_expr}},\code{\link[rlang]{missing_arg}}
\code{\link[purrr]{map}},\code{\link[purrr]{set_names}}

Other ff functions: 
\code{\link{brake_ff}()}
}
\concept{ff functions}
\concept{flag functions}
