% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grant.R
\name{grant_all_in_schema}
\alias{grant_all_in_schema}
\title{Grant All Privileges on all Tables}
\usage{
grant_all_in_schema(
  conn,
  schema,
  users,
  groups,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  ...
)
}
\arguments{
\item{users}{Usernames}

\item{groups}{Group names}
}
\description{
Grant all privileges to a schema to either a group or a user.
}
