% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{grantSchema}
\alias{grantSchema}
\title{Grant All Privileges to a Schema}
\usage{
grantSchema(conn, schema, user = NULL, group = NULL)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION}

\item{schema}{PARAM_DESCRIPTION}

\item{user}{PARAM_DESCRIPTION, Default: NULL}

\item{group}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Grant all privileges to a schema to either a group or a user.
}
