% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join2.R
\name{ij2}
\alias{ij2}
\title{Second Generation Inner Join}
\usage{
ij2(
  data,
  omopSchema,
  omopTable,
  ...,
  write_schema,
  cast_to_varchar = TRUE,
  case_insensitive = TRUE,
  render_sql = TRUE,
  verbose = TRUE
)
}
\description{
Second Generation Inner Join
}
