% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_plr_installed.R
\name{is_plr_installed}
\alias{is_plr_installed}
\title{Is PLR Installed?}
\usage{
is_plr_installed(
  conn,
  conn_fun = "pg13::local_connect()",
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  checks = ""
)
}
\description{
Is PLR Installed?
}
\seealso{
Other PLR: 
\code{\link{cat_plr_example}()}
}
\concept{PLR}
