% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join2.R
\name{join2}
\alias{join2}
\title{2nd Generation Join}
\usage{
join2(
  conn,
  write_schema,
  kind = c("left", "right", "inner", "full"),
  data,
  schema,
  table_name,
  ...,
  cast_to_varchar = TRUE,
  case_insensitive = TRUE,
  verbose = TRUE,
  render_sql = TRUE
)
}
\description{
2nd Generation Join
}
