% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_users.R
\name{list_users}
\alias{list_users}
\title{List Users}
\usage{
list_users(
  conn,
  conn_fun,
  verbose = TRUE,
  render_sql = TRUE,
  warn_no_rows = TRUE,
  render_only = FALSE,
  ...
)
}
\description{
List Users
}
