% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{lj2}
\alias{lj2}
\title{Second Generation Left Join}
\usage{
lj2(
  data,
  omopSchema,
  omopTable,
  ...,
  writeSchema,
  cast_to_varchar = TRUE,
  case_insensitive = TRUE,
  render_sql = TRUE,
  verbose = TRUE
)
}
\description{
Second Generation Left Join
}
