% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{loadCachedJoin}
\alias{loadCachedJoin}
\title{Load a Cached Join Resultset}
\usage{
loadCachedJoin(..., db, schema)
}
\arguments{
\item{...}{Arguments passed to the buildJoinQuery function}

\item{db}{name of database}

\item{schema}{name of schema}
}
\description{
This function will load the cached resultset from a SQL query created by the buildJoinQuery function to a subdirectory of the cache root directory, the path of which will be {db/schema}.
}
\keyword{internal}
