% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_funs.R
\name{loadCachedQuery}
\alias{loadCachedQuery}
\title{Load a Cached Query Resultset}
\usage{
loadCachedQuery(sqlQuery, db)
}
\arguments{
\item{sqlQuery}{sql statement of length 1 that will be the sole key for cache retrieval.}

\item{db}{name of database}

\item{schema}{name of schema}
}
\description{
This function will load the cached resultset from a SQL query to a subdirectory of the cache root directory, the path of which will be {db/schema}.
}
