% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_conn.R
\name{local_conn}
\alias{local_conn}
\title{Create a conn object on a local database}
\usage{
local_conn(dbname, schema = NULL)
}
\arguments{
\item{dbname}{name of database}

\item{schema}{optional and applies to postgres only if used.}
}
\description{
If the database is for the local UMLS, a MySQL connection is made. Otherwise, the connect is made as a local postgres database.
}
