% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{local_file_query}
\alias{local_file_query}
\title{Query local Postgess from a file}
\usage{
local_file_query(file, dbname = "athena", port = 5432)
}
\description{
Query local Postgess from a file
}
\keyword{internal}
