% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decamel-deprec.R
\name{local_file_send}
\alias{local_file_send}
\title{Send Statement to a local Postgess from a file}
\usage{
local_file_send(file, dbname = "athena", port = 5432)
}
\description{
Send Statement to a local Postgess from a file
}
