% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decamel-deprec.R
\name{local_send}
\alias{local_send}
\title{Send a SQL Statement to a Local Postgres}
\usage{
local_send(sql_statement, dbname = "athena", port = 5432)
}
\description{
Send a SQL Statement to a Local Postgres
}
