% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_db.R
\name{localConnect}
\alias{localConnect}
\title{Connect to Local PostGres Server}
\usage{
localConnect(dbname = "athena", port = 5432)
}
\arguments{
\item{dbname}{database name with the corresponding 'localhost/{dbname}' server}

\item{port}{defults to "5432"}

\item{schema}{defaults to NULL}
}
\description{
Connect to Local PostGres Server
}
