% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{localFileQuery}
\alias{localFileQuery}
\title{Query local Postgess from a file}
\usage{
localFileQuery(file, dbname = "athena", port = 5432)
}
\description{
Query local Postgess from a file
}
\keyword{internal}
