% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localFileSend.R
\name{localFileSend}
\alias{localFileSend}
\title{Send Statement to a local Postgess from a file}
\usage{
localFileSend(file, dbname = "athena", port = 5432)
}
\description{
Send Statement to a local Postgess from a file
}
