% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{localSend}
\alias{localSend}
\title{Send a SQL Statement to a Local Postgres}
\usage{
localSend(sql_statement, dbname = "athena", port = 5432)
}
\description{
Send a SQL Statement to a Local Postgres
}
\keyword{internal}
