% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_stop_ff}
\alias{log_stop_ff}
\title{Log a Stop Function Factory}
\usage{
log_stop_ff(datetime_field = "log_datetime", schema, log_table)
}
\arguments{
\item{datetime_field}{Name of the datetime field. Defaults to \code{log_datetime}.}

\item{log_table}{Log table name}
}
\value{
A function that logs the stop of an activity.
}
\description{
Create a function that will log the stop of an activity to the provided \verb{Log Table}.
}
\seealso{
\code{\link[tibble]{tibble}}
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}},\code{\link[dplyr]{reexports}}

Other logging functions: 
\code{\link{create_log}()},
\code{\link{log_error_ff}()},
\code{\link{log_start_ff}()}
}
\concept{logging functions}
