% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-typewrite.R
\name{lowLevel_typewrite_activity}
\alias{lowLevel_typewrite_activity}
\title{Typewrite An Activity}
\usage{
lowLevel_typewrite_activity(activity)
}
\arguments{
\item{activity}{Character string.}
}
\description{
Return the provided text in yellow along with a
timestamp.
}
\seealso{
\code{\link[secretary]{c("typewrite", "yellowTxt")}}
}
