% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-typewrite.R
\name{lowLevel_typewrite_alert_danger}
\alias{lowLevel_typewrite_alert_danger}
\title{Typewrite Alert Danger}
\usage{
lowLevel_typewrite_alert_danger(text)
}
\arguments{
\item{text}{Character string.}
}
\description{
Return a timestamped console message in bold red with
exclamation mark.
}
\seealso{
\code{\link[secretary]{c("typewrite", "redTxt", "enbold)}}
}
