% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-typewrite.R
\name{lowLevel_typewrite_sql}
\alias{lowLevel_typewrite_sql}
\title{Typewrite SQL}
\usage{
lowLevel_typewrite_sql(sql_statement, style = c("inline", "chunk"))
}
\arguments{
\item{sql_statement}{SQL statement.}

\item{style}{The \emph{inline} style where the statement is flattened
into a single line or a \emph{chunk} style where the statement is
taken in its native format and returned within a set of
ticks for R Markdown, Default: "inline".}
}
\description{
Wrapper around \code{cat} function in R that returns a
timestamped \verb{SQL:} heading followed by the SQL statement.
}
\seealso{
\code{\link[stringr]{str_replace}}
\code{\link[secretary]{c("typewrite", "enbold", "blueTxt")}}
}
