% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{ls_db}
\alias{ls_db}
\title{List Databases}
\usage{
ls_db(conn, verbose = TRUE, render_sql = TRUE, render_only = FALSE)
}
\description{
List all the databases in a connection
}
\seealso{
Other list functions: 
\code{\link{lsFields}()},
\code{\link{lsSchema}()},
\code{\link{lsTables}()},
\code{\link{ls_fields}()},
\code{\link{ls_schema}()},
\code{\link{ls_tables}()},
\code{\link{ls_users}()}
}
\concept{list functions}
