% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_extensions.R
\name{ls_extensions}
\alias{ls_extensions}
\title{List Extensions}
\usage{
ls_extensions(
  conn,
  conn_fun = "pg13::local_connect()",
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  checks = ""
)
}
\value{
Vector of extension names
}
\description{
List all the installed extensions.
}
