% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{ls_tables}
\alias{ls_tables}
\title{List Tables}
\usage{
ls_tables(conn, schema, verbose = TRUE, render_sql = TRUE)
}
\arguments{
\item{conn}{Connection object}

\item{schema}{The target schema for the operation.}

\item{verbose}{If TRUE, details on the activity are returned in the console, such as when the querying starting and was completed.}

\item{render_sql}{If TRUE, the SQL statement for the operation is returned in the console.}
}
\description{
List Tables
}
\seealso{
Other list functions: 
\code{\link{ls_fields}()},
\code{\link{ls_schema}()}
}
\concept{list functions}
