% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{ls_users}
\alias{ls_users}
\title{List Users}
\usage{
ls_users(
  conn,
  conn_fun,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  ...
)
}
\description{
List Users
}
\seealso{
Other list functions: 
\code{\link{lsFields}()},
\code{\link{lsSchema}()},
\code{\link{lsTables}()},
\code{\link{ls_db}()},
\code{\link{ls_fields}()},
\code{\link{ls_schema}()},
\code{\link{ls_tables}()}
}
\concept{list functions}
