% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{lsTables}
\alias{lsTables}
\title{List Tables}
\usage{
lsTables(conn, schema, verbose = TRUE, render_sql = TRUE)
}
\description{
List Tables
}
\seealso{
Other list functions: 
\code{\link{lsFields}()},
\code{\link{lsSchema}()},
\code{\link{ls_db}()},
\code{\link{ls_fields}()},
\code{\link{ls_schema}()},
\code{\link{ls_tables}()},
\code{\link{ls_users}()}
}
\concept{list functions}
