% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate.R
\name{migrate}
\alias{migrate}
\title{Migrate Tables}
\usage{
migrate(
  conn_1,
  schema_1,
  ...,
  conn_2,
  schema_2,
  drop_existing,
  verbose = TRUE,
  render_sql = TRUE
)
}
\arguments{
\item{...}{(optional) Character strings of names of tables to migrate.}
}
\description{
Migrate Tables in one schema to another schema that does or does not belong to
another connection. The new schema can be of a different name, but the table names
remain the same.
}
\seealso{
\code{\link[rlang]{list2}}
\code{\link[secretary]{typewrite_warning}},\code{\link[secretary]{c("typewrite", "typewrite")}},\code{\link[secretary]{typewrite_progress}},\code{\link[secretary]{character(0)}}
\code{\link[purrr]{map}}
}
