% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn-env.R
\name{open_conn}
\alias{open_conn}
\title{FUNCTION_TITLE}
\usage{
open_conn(conn_fun, conn_name)
}
\arguments{
\item{conn_fun}{Function call for the connection as a character string.}

\item{conn_name}{Name of the connection object in the \code{pg13_connection_env} environment}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[secretary]{c("typewrite", "typewrite")}},\code{\link[secretary]{italicize}}
\code{\link[cli]{cli_alert}}
\code{\link[rlang]{parse_expr}}
}
