% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseSQL.R
\name{parseSQL}
\alias{parseSQL}
\title{Parse SQL into Single Statements}
\usage{
parseSQL(sql_statement)
}
\value{
List of sql statements
}
\description{
This function is helpful in splitting a SQL file into digestible smaller statements based on the semicolon for troubleshooting or to skip of the statements that do not work within the full input SQL.
}
