% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{peek}
\alias{peek}
\title{Peek}
\usage{
peek(
  conn,
  conn_fun,
  schema,
  table,
  limit = 5,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE
)
}
\arguments{
\item{limit}{Number of rows to randomly return. Default: 5.}
}
\description{
Get a quick glance of a handful of rows in a table.
}
\seealso{
\code{\link[SqlRender]{render}}
}
