% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn-env.R
\docType{class}
\name{pg_env-class}
\alias{pg_env-class}
\alias{pg13_env}
\title{PG13 Environment S4 Class}
\description{
This an optional environment class that stores all the connections using this package.
}
