% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn-env.R
\name{[[<-,pg_env,character,missing,ANY-method}
\alias{[[<-,pg_env,character,missing,ANY-method}
\title{Methods: PG13 environment}
\usage{
\S4method{[[}{pg_env,character,missing,ANY}(x, i) <- value
}
\description{
Methods: PG13 environment
}
