% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgQuery.R
\name{pgQuery}
\alias{pgQuery}
\title{Send Query a Postgres connection}
\usage{
pgQuery(conn, sql_statement, ...)
}
\arguments{
\item{conn}{Connection object}

\item{sql_statement}{SQL Query}

\item{...}{Additional arguments to pass onto the DatabaseConnector::dbGetQuery function}
}
\description{
Send Query a Postgres connection
}
