% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgSend.R
\name{pgSend}
\alias{pgSend}
\title{Send a SQL Statement to Postgres}
\usage{
pgSend(conn, sql_statement, ...)
}
\arguments{
\item{conn}{Connection object}

\item{sql_statement}{SQL to send}

\item{...}{Additional arguments to pass onto the DatabaseConnector::dbSendStatement function}
}
\description{
Send a SQL Statement to Postgres
}
