% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decamel-utils1.R, R/utils1.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs

lhs \%>\% rhs
}
\description{
See \code{magrittr::\link[magrittr:pipe]{\%>\%}} for details.

See \code{magrittr::\link[magrittr:pipe]{\%>\%}} for details.
}
\keyword{internal}
