% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_vector.R
\name{prepare_vector}
\alias{prepare_vector}
\title{Convert vector to arguments}
\usage{
prepare_vector(vector)
}
\arguments{
\item{vector}{vector of arguments}
}
\description{
Convert vector to arguments
}
