% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FF.R
\name{q_creds_ff}
\alias{q_creds_ff}
\title{Query Function Factory}
\usage{
q_creds_ff(user, password, port, server)
}
\value{
Query function
}
\description{
Write a query function that automatically connects and disconnects on exit based on the connection details parameters given.
}
