% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_ff.R
\name{q_ff}
\alias{q_ff}
\title{Query Function Factory}
\usage{
q_ff(user, password, port, server, cache_dirs = NULL)
}
\value{
Query function
}
\description{
Write a query function that automatically connects and disconnects on exit based on the connection details parameters given.
}
