% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{query}
\alias{query}
\title{Send Query to any Postgres connection}
\usage{
query(
  conn,
  conn_fun,
  sql_statement,
  verbose = TRUE,
  render_sql = TRUE,
  warn_no_rows = TRUE,
  render_only = FALSE,
  ...
)
}
\arguments{
\item{conn}{Connection object}

\item{sql_statement}{SQL Query}

\item{warn_no_rows}{If TRUE, a warning is given that query has returned 0 rows.}

\item{render_only}{If TRUE, \code{sql_statement} is printed to the console, but the query is not run. This is an option to QA SQL if a query is throwing errors.}

\item{...}{Additional arguments to pass onto \code{\link[DatabaseConnector]{dbGetQuery}}}
}
\description{
Send Query to any Postgres connection
}
