% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_postgres_from_file.R
\name{query_pg_from_file}
\alias{query_pg_from_file}
\title{Query a local postgres database from a file}
\usage{
query_pg_from_file(path_to_sql_file, dbname, schema = NULL, port = "5432")
}
\arguments{
\item{path_to_sql_file}{path to .sql file}

\item{dbname}{name of local postgres database}
}
\description{
Query a local postgres database from a file
}
