% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_funs.R
\name{queryList}
\alias{queryList}
\title{Query a List of SQL Statements}
\usage{
queryList(conn, sqlList, verbose = TRUE, progressBar = TRUE, ...)
}
\arguments{
\item{conn}{Connection object}

\item{sqlList}{list object of queries}

\item{...}{Additional arguments to pass onto the DatabaseConnector::dbGetQuery function}
}
\description{
Iteratively query over a list of SQL Statements such as the object returned by \code{\link{parseSQL}}.
}
