% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{quietly.conn_db}
\alias{quietly.conn_db}
\title{Connect without Console Messages}
\usage{
quietly.conn_db(...)
}
\description{
Connect without Console Messages
}
\seealso{
Other connection functions: 
\code{\link{conn_db}()},
\code{\link{connect_ff}()},
\code{\link{connect}()},
\code{\link{dc0}()},
\code{\link{dc}()}
}
\concept{connection functions}
\keyword{internal}
