% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{read_table}
\alias{read_table}
\title{Read table in the given database}
\usage{
read_table(table_name, dbname, schema)
}
\arguments{
\item{table_name}{name of new table}

\item{dbname}{name of database for new table}

\item{schema}{name of schema for new table}
}
\description{
Read table in the given database
}
