% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{read_view}
\alias{read_view}
\title{Read View}
\usage{
read_view(conn, conn_fun, schema, viewName, verbose = TRUE, render_sql = TRUE)
}
\description{
Read View
}
\seealso{
Other read functions: 
\code{\link{read_table}()}
}
\concept{read functions}
\concept{view functions}
