% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{readTable}
\alias{readTable}
\title{Read an Entire Table}
\usage{
readTable(conn, schema, tableName, verbose = TRUE, render_sql = TRUE)
}
\description{
Shortcut for a \verb{SELECT *} SQL statement.
}
\keyword{internal}
