% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{refresh_mat_view}
\alias{refresh_mat_view}
\title{Refresh a Materialized View}
\usage{
refresh_mat_view(conn, schema, matViewName, verbose = TRUE, render_sql = TRUE)
}
\description{
Refresh a Materialized View
}
\concept{materialized view functions}
