% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{refresh_mview}
\alias{refresh_mview}
\title{Refresh a Materialized View}
\usage{
refresh_mview(conn, schema, mview, verbose = TRUE, render_sql = TRUE)
}
\description{
Refresh a Materialized View
}
\concept{materialized view functions}
