% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{refresh_table}
\alias{refresh_table}
\title{Refresh Table with New Data}
\usage{
refresh_table(conn, schema, tableName, .data)
}
\description{
A refresh is when a table needs to be overwritten and the table that is being overwritten is off-loaded to another table with today's date and the index of iterations for that day.
}
\keyword{internal}
