% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec.R
\name{refreshTable}
\alias{refreshTable}
\title{Refresh Table with New Data}
\usage{
refreshTable(conn, schema, tableName, .data)
}
\description{
A refresh is when a table needs to be overwritten and the table that is being overwritten is off-loaded to another table with today's date and the index of iterations for that day.
}
