% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decamel-render.R
\name{render_create_schema}
\alias{render_create_schema}
\title{Render SQL to Create a Schema}
\usage{
render_create_schema(schema)
}
\description{
Renders a SQL statement as a string that creates a schema.
}
