% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_drop_schema}
\alias{render_drop_schema}
\title{Render SQL to Drop a Schema}
\usage{
render_drop_schema(schema, cascade = FALSE, if_exists = TRUE)
}
\arguments{
\item{cascade}{If TRUE, a DROP SCHEMA CASCADE is performed.}
}
\description{
Drop a schema if it exists
}
