% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decamel-render.R
\name{render_rename_db}
\alias{render_rename_db}
\title{Render SQL to Rename a Table}
\usage{
render_rename_db(schema, db, newDB)
}
\description{
This will rename a table within a schema, but not move the table out of a schema.
}
