% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_rename_table}
\alias{render_rename_table}
\title{Render SQL to Rename a Table}
\usage{
render_rename_table(schema, tableName, newTableName)
}
\description{
This will rename a table within a schema, but not move the table out of a schema.
}
