% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{renderDropSchema}
\alias{renderDropSchema}
\title{Render SQL to Drop a Schema}
\usage{
renderDropSchema(schema, cascade = FALSE, if_exists = TRUE)
}
\arguments{
\item{cascade}{If TRUE, a DROP SCHEMA CASCADE is performed.}
}
\description{
Drop a schema if it exists
}
