% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{renderInfoSchemaCols}
\alias{renderInfoSchemaCols}
\title{Render SQL that returns Column Information}
\usage{
renderInfoSchemaCols(schema)
}
\arguments{
\item{schema}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[SqlRender]{render}}
}
