% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{renderRenameDB}
\alias{renderRenameDB}
\title{Render SQL to Rename a Table}
\usage{
renderRenameDB(schema, db, newDB)
}
\description{
This will rename a table within a schema, but not move the table out of a schema.
}
