% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderRenameTable.R
\name{renderRenameTable}
\alias{renderRenameTable}
\title{Render SQL to Rename a Table}
\usage{
renderRenameTable(schema, tableName, newTableName)
}
\description{
This will rename a table within a schema, but not move the table out of a schema.
}
