% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{renderRowCount}
\alias{renderRowCount}
\title{Render SQL for a Table Row Count}
\usage{
renderRowCount(fields = "*", distinct = FALSE, schema, tableName)
}
\arguments{
\item{fields}{Fields selected for. Defaults to "*".}

\item{distinct}{If TRUE, the distinct row count will be returned.}

\item{schema}{If NULL, defaults to "public"}
}
\description{
Render SQL for a Table Row Count
}
