% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R, R/deprecated.R
\name{rm_if_closed}
\alias{rm_if_closed}
\title{Remove a Closed Connection Object}
\usage{
rm_if_closed(conn)

rm_if_closed(conn)

rm_if_closed(conn)
}
\arguments{
\item{conn}{Postgres connection object}
}
\description{
If a connection is not open, it is removed from the parent environment from which this function is called.

(Deprecated) This function removes a connection object from the Global Environment if it is closed.

If a connection is not open, it is removed from the parent environment from which this function is called.
}
\keyword{internal}
