% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec.R
\name{rmIfClosed}
\alias{rmIfClosed}
\title{Remove a Closed Connection Object}
\usage{
rmIfClosed(conn)
}
\arguments{
\item{conn}{Postgres connection object}
}
\description{
(Deprecated) This function removes a connection object from the Global Environment if it is closed.
}
