% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{save_sql}
\alias{save_sql}
\title{Export a sql statement to a file}
\usage{
save_sql(sql_statement, file, append = TRUE, ...)

save_sql(sql_statement, file, append = TRUE, ...)
}
\arguments{
\item{sql_statement}{sql statement R object}

\item{file}{File to write to.}

\item{...}{Additional arguments passed to the readr::write_lines function}
}
\description{
Export a sql statement to a file

Export a sql statement to a file
}
