% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/exists.R
\name{schema_exists}
\alias{schema_exists}
\title{Does a schema exist?}
\usage{
schema_exists(conn, schema)

schema_exists(conn, schema)
}
\description{
Logical that checks if a schema exists in the database. The \code{schema} argument is in formatted in all lowercase prior to checking against what is present in the database.

Logical that checks if a schema exists in the database. The \code{schema} argument is in formatted in all lowercase prior to checking against what is present in the database.
}
\seealso{
Other logical functions: 
\code{\link{db_exists}()},
\code{\link{field_exists}()},
\code{\link{table_exists}()}

Other logical functions: 
\code{\link{db_exists}()},
\code{\link{field_exists}()},
\code{\link{table_exists}()}
}
\concept{logical functions}
