% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{send}
\alias{send}
\title{Send a SQL Statement to Postgres}
\usage{
send(
  conn,
  conn_fun,
  sql_statement,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  ...
)
}
\description{
Send a SQL Statement to Postgres
}
