% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{send_ff}
\alias{send_ff}
\title{Send Function Factor}
\usage{
send_ff(user, password, port, server)

send_ff(user, password, port, server)
}
\description{
Send Function Factor

Send Function Factor
}
\keyword{internal}
