% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/utils2.R
\name{source_file_path}
\alias{source_file_path}
\title{Get SourceFile Path}
\usage{
source_file_path(instSubdir, FileName, package)

source_file_path(instSubdir, FileName, package)
}
\arguments{
\item{instSubdir}{Name of subdirectory in the inst/ folder}

\item{FileName}{Name of file in subdirectory}

\item{package}{Package name}
}
\description{
This function provides the path for files installed within a given package's library.

This function provides the path for files installed within a given package's library.
}
