% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceFilePath.R
\name{sourceFilePath}
\alias{sourceFilePath}
\title{Get SourceFile Path}
\usage{
sourceFilePath(instSubdir, FileName, package)
}
\arguments{
\item{instSubdir}{Name of subdirectory in the inst/ folder}

\item{FileName}{Name of file in subdirectory}

\item{package}{Package name}
}
\description{
This function provides the path for files installed within a given package's library.
}
