% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_funs.R
\name{summarize_db}
\alias{summarize_db}
\title{FUNCTION_TITLE}
\usage{
summarize_db(conn, output.var, verbose = TRUE, progressBar = TRUE)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION}

\item{output.var}{PARAM_DESCRIPTION}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}

\item{progressBar}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}}
\code{\link[dplyr]{bind}},\code{\link[dplyr]{select}},\code{\link[dplyr]{distinct}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{reexports}}
\code{\link[progress]{progress_bar}}
\code{\link[tidyr]{pivot_wider}}
\code{\link[secretary]{typewrite_bold}}
}
