% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decamel-schema_funs.R
\name{summarize_schema}
\alias{summarize_schema}
\title{Summarize a Schema}
\usage{
summarize_schema(conn, schema, resultTableName)
}
\arguments{
\item{conn}{Postgres connection}

\item{schema}{Schema to summarize}

\item{resultTableName}{Table that the final output will be written to.}
}
\value{
New Table in the same schema
}
\description{
Summarize a Schema
}
\seealso{
\code{\link[purrr]{map}}
\code{\link[tibble]{as_tibble}}
\code{\link[dplyr]{bind}},\code{\link[dplyr]{mutate}}
}
