% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{summarizeSchema}
\alias{summarizeSchema}
\title{Summarize a Schema}
\usage{
summarizeSchema(conn, schema, resultTableName)
}
\arguments{
\item{conn}{Postgres connection}

\item{schema}{Schema to summarize}

\item{resultTableName}{Table that the final output will be written to.}
}
\value{
New Table in the same schema
}
\description{
Summarize a Schema
}
\seealso{
\code{\link[purrr]{map}}
\code{\link[tibble]{as_tibble}}
\code{\link[dplyr]{bind}},\code{\link[dplyr]{mutate}}
}
\keyword{internal}
