% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/draft.R
\name{table.path}
\alias{table.path}
\title{Construct schemaTableName}
\usage{
table.path(schema, table_name)

table.path(schema, table_name)
}
\description{
construct schemaTableName from the schema and tableName

construct schemaTableName from the schema and tableName
}
