% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/list.R
\name{table_exists}
\alias{table_exists}
\title{Does a table exist?}
\usage{
table_exists(conn, schema, tableName)

table_exists(conn, schema, tableName)
}
\description{
Does a table exist?

Does a table exist?
}
\seealso{
Other logical functions: 
\code{\link{field_exists}()},
\code{\link{schema_exists}()}

Other logical functions: 
\code{\link{field_exists}()},
\code{\link{schema_exists}()}
}
\concept{logical functions}
