% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build2.R, R/deprecated.R
\name{terminate_build}
\alias{terminate_build}
\title{Terminate a SQL Statement with a semicolon}
\usage{
terminate_build(sql_statement)

terminate_build(sql_statement)
}
\description{
Terminate a SQL Statement with a semicolon

Terminate a SQL Statement with a semicolon
}
\keyword{internal}
