% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{terminateBuild}
\alias{terminateBuild}
\title{Terminate a SQL Statement with a semicolon}
\usage{
terminateBuild(sql_statement)

terminateBuild(sql_statement)
}
\description{
Terminate a SQL Statement with a semicolon

Terminate a SQL Statement with a semicolon
}
