% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-typewrite.R
\name{typewrite_alert_success}
\alias{typewrite_alert_success}
\title{Typewrite Alert Success}
\usage{
typewrite_alert_success(text)
}
\arguments{
\item{text}{Character string.}
}
\description{
Return a message in a bold green with a checkmark.
}
\seealso{
\code{\link[secretary]{c("typewrite", "greenTxt", "enbold)}}
}
