% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_timeout.R
\name{with_timeout}
\alias{with_timeout}
\title{Wrap to include a timeout to evaluation.}
\usage{
with_timeout(expr, cpu, elapsed)
}
\arguments{
\item{expr}{Expression to evaluate.}

\item{cpu}{double (of length one).  Set a limit on
    the total or elapsed cpu time in seconds, respectively.}

\item{elapsed}{double (of length one).  Set a limit on
    the total or elapsed cpu time in seconds, respectively.}
}
\description{
Wrap to include a timeout to evaluation.
}
