% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sql_where_in.R
\name{write_query_where_in}
\alias{write_query_where_in}
\title{Writes a SQL query that includes a WHERE IN clause based on a vector}
\usage{
write_query_where_in(
  select = "*",
  table_name,
  column_name,
  where_in_vector,
  limit = NULL
)
}
\arguments{
\item{select}{vector of table_column_names. Defaults to "*".}

\item{table_name}{name of table}

\item{column_name}{name of column for WHERE IN clause}

\item{where_in_vector}{vector of WHERE IN clause}

\item{limit}{limit count}
}
\description{
Writes a SQL query that includes a WHERE IN clause based on a vector
}
