% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{write_sql_file}
\alias{write_sql_file}
\title{Export a sql statement to a file}
\usage{
write_sql_file(sql_statement, file, append = TRUE, ...)
}
\arguments{
\item{sql_statement}{sql statement R object}

\item{file}{File to write to.}

\item{...}{Additional arguments passed to the readr::write_lines function}
}
\description{
Export a sql statement to a file
}
