% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sql_for_distinct_counts.R
\name{write_sql_for_distinct_counts}
\alias{write_sql_for_distinct_counts}
\title{Write sql that gets distinct counts}
\usage{
write_sql_for_distinct_counts(table_name, select = "*")
}
\arguments{
\item{table_name}{name of table with schema if applicable}

\item{select}{character vector of column names to call the DISTINCT function on}
}
\value{
string of length one of a complete sql statement that will return distinct counts when executed
}
\description{
Write sql that gets distinct counts
}
