% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sql_for_random_records.R
\name{write_sql_for_random_records}
\alias{write_sql_for_random_records}
\title{Writes SQL to get random records of a certain limit}
\usage{
write_sql_for_random_records(table_name, limit = 1000, select = "*")
}
\arguments{
\item{table_name}{name of table, including schema if applicable}

\item{limit}{value LIMIT is assigned to in the sql statement}

\item{select}{character vector of column names for SELECT statement}
}
\value{
character string of length one of a sql statement that queries a table for a desired number of random records
}
\description{
Writes SQL to get random records of a certain limit
}
