% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildQueryString.R
\name{write_sql_query_string_as_vector}
\alias{write_sql_query_string_as_vector}
\title{Write a SQL statement that loops over a word in a string}
\usage{
write_sql_query_string_as_vector(
  string,
  split,
  select = "*",
  table_name,
  column_name,
  case_insensitive = TRUE,
  limit = NULL
)
}
\description{
Write a SQL statement that loops over a word in a string
}
