% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sql_statement_to_load_rrf.R
\name{write_sql_statement_to_load_rrf}
\alias{write_sql_statement_to_load_rrf}
\title{Return SQL statements into the console that loads RRF in raw format into a corresponding table}
\usage{
write_sql_statement_to_load_rrf(rrf_filenames)
}
\arguments{
\item{rrf_filenames}{character vector of rrf filenames that will be loaded into tables of the same basename}
}
\description{
Return SQL statements into the console that loads RRF in raw format into a corresponding table
}
