% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sql_to_copy_from_csv.R
\name{write_sql_to_copy_from_csv}
\alias{write_sql_to_copy_from_csv}
\title{Writes a dataframe as a table to the given database}
\usage{
write_sql_to_copy_from_csv(csv_fn, to_table_name)
}
\description{
Writes a dataframe as a table to the given database
}
