% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sql_to_create_schema.R
\name{write_sql_to_create_schema}
\alias{write_sql_to_create_schema}
\title{Write SQL to create a schema}
\usage{
write_sql_to_create_schema(schema)
}
\description{
Write SQL to create a schema
}
