% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sql_to_drop_create_table_from_df.R
\name{write_sql_to_drop_create_table_from_df}
\alias{write_sql_to_drop_create_table_from_df}
\title{Writes a DROP TABLE and CREATE TABLE sql statement for a given dataframe}
\usage{
write_sql_to_drop_create_table_from_df(
  dataframe,
  to_table_name,
  add_to_varchar = 10
)
}
\arguments{
\item{dataframe}{dataframe to be loaded into table}

\item{to_table_name}{name of the table that the dataframe will be loaded into}

\item{add_to_varchar}{integer that is added to the character count for good measure.}
}
\description{
Writes a DROP TABLE and CREATE TABLE sql statement for a given dataframe
}
