% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_staging_table.R
\name{write_staging_table}
\alias{write_staging_table}
\title{Write a Staging Table.}
\usage{
write_staging_table(
  conn,
  conn_fun,
  schema,
  data,
  drop_existing = FALSE,
  verbose = TRUE,
  render_sql = TRUE,
  ...
)
}
\arguments{
\item{conn}{Connection object}

\item{schema}{The target schema for the operation.}

\item{data}{A dataframe or tibble.}

\item{verbose}{If TRUE, details on the activity are returned in the console, such as when the querying starting and was completed.}

\item{render_sql}{If TRUE, the SQL statement for the operation is returned in the console.}
}
\description{
A `Staging Table`` is one that is named automatically in the "V{Timestamp}" format and is dropped on exit in the parent frame from which the function is being called.
}
