% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{write_table2}
\alias{write_table2}
\title{Write a Table}
\usage{
write_table2(conn = conn, schema, tableName, .data, ...)
}
\arguments{
\item{...}{Additional arguments passed to DatabaseConnector::dbWriteTable}
}
\description{
Unlike the dropTable and renameTable functions, this function is a wrapper around the DatabaseConnector::dbWriteTable function rather than one where a SQL statement is rendered using the SqlRender package. This function that converts all inputs to the data.frame class, especially in cases where the input is a tibble, in which case an error would be thrown when writing.
}
\keyword{internal}
