% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_where_in_string.R
\name{write_where_in_string}
\alias{write_where_in_string}
\title{Take a vector and create a string for WHERE IN SQL}
\usage{
write_where_in_string(where_in_vector, as_character = TRUE, formatted = FALSE)
}
\arguments{
\item{where_in_vector}{vector that will filter the resultset}

\item{as_character}{If TRUE, includes single quotes around each value. If FALSE, the vector is printed as-is.}
}
\description{
This function takes a vector of 1 or more values and creates a bracketed, comma-separated string of length 1 that can be inserted into the SQL statement for filtering.
}
