% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{writeStagingTable}
\alias{writeStagingTable}
\title{Write a Table that is Dropped On Exit}
\usage{
writeStagingTable(
  conn,
  conn_fun,
  schema,
  tableName,
  data,
  drop_existing = FALSE,
  verbose = TRUE,
  render_sql = TRUE,
  ...
)
}
\description{
A Staging Table is one that is dropped on exit in the parent frame from which the function is being called.
}
\keyword{internal}
