% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_join_query}
\alias{build_join_query}
\title{Build a Query for a Join}
\usage{
build_join_query(
  select_table_fields = "*",
  select_join_on_fields = "*",
  distinct = FALSE,
  schema,
  table,
  column,
  join_on_schema,
  join_on_table,
  join_on_column,
  kind = c("LEFT", "RIGHT", "INNER", "FULL"),
  where_in_field,
  where_in_vector,
  where_in_join_on_field,
  where_in_join_on_vector,
  where_not_in_field,
  where_not_in_vector,
  where_not_in_join_on_field,
  where_not_in_join_on_vector,
  where_is_null_field,
  where_is_not_null_field,
  where_is_null_join_on_field,
  where_is_not_null_join_on_field,
  case_insensitive,
  limit,
  random
)
}
\description{
Build a query the joins 2 tables, respectively aliased as \code{a} and \code{b} in the query itself.
}
\seealso{
\code{\link[cli]{cli_alert}}
\code{\link[SqlRender]{render}}
}
