% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decamel-build1.R, R/decamel-decamel-build1.R
\name{build_join_query}
\alias{build_join_query}
\title{Build a JOIN SQL Query}
\usage{
build_join_query(
  fields = "*",
  distinct = FALSE,
  schema,
  tableName,
  column,
  joinOnSchema,
  joinOnTableName,
  joinOnColumn,
  joinType = "LEFT",
  whereInField = NULL,
  whereInVector = NULL,
  whereNotInField = NULL,
  whereNotInVector = NULL,
  caseInsensitive = TRUE,
  n = NULL,
  n_type = c("limit", "random")
)

build_join_query(
  fields = "*",
  distinct = FALSE,
  schema,
  tableName,
  column,
  joinOnSchema,
  joinOnTableName,
  joinOnColumn,
  joinType = "LEFT",
  whereInField = NULL,
  whereInVector = NULL,
  whereNotInField = NULL,
  whereNotInVector = NULL,
  caseInsensitive = TRUE,
  n = NULL,
  n_type = c("limit", "random")
)
}
\value{
SQL statement as a character string.

SQL statement as a character string.
}
\description{
A SQL query is built using the given arguments. Currently, only 1 whereIn and whereNot in parameters can be set.

A SQL query is built using the given arguments. Currently, only 1 whereIn and whereNot in parameters can be set.
}
