% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_query}
\alias{build_query}
\title{Build a Simple SQL Query}
\usage{
build_query(
  fields = "*",
  distinct = FALSE,
  schema,
  table,
  where_in_field,
  where_in_vector,
  where_not_in_field,
  where_not_in_vector,
  where_is_null_field,
  where_is_not_null_field,
  case_insensitive = TRUE,
  limit,
  random
)
}
\description{
Build a Simple SQL Query
}
\examples{
# Building SQL Statements using R
library(pg13)

# Build a Simple Query
build_query(schema = "public",
            table = "concept_synonym")

build_query(distinct = TRUE,
            schema = "public",
            table = "concept_synonym")


build_query(fields = c("concept_id", "concept_name"),
            distinct = TRUE,
            schema = "public",
            table = "concept_synonym")

# WHERE filters may be applied, but will be ignored if the field-value pair is not provided
build_query(fields = c("concept_id", "concept_name"),
            schema = "public",
            table = "concept_synonym",
            where_in_field = "concept_synonym")

# The query defaults to transforming elements to lowercase for case insensitivity that can be turned off
build_query(fields = c("concept_id", "concept_name"),
            schema = "public",
            table = "concept_synonym",
            where_in_field = "concept_synonym",
            where_in_vector = c("Myocardial Infarction", "Heart Attack"))

build_query(fields = c("concept_id", "concept_name"),
            schema = "public",
            table = "concept_synonym",
            where_in_field = "concept_synonym",
            where_in_vector = c("Myocardial Infarction", "Heart Attack"),
            case_insensitive = FALSE)

}
\seealso{
\code{\link[cli]{cli_alert}}
\code{\link[SqlRender]{render}}
}
