% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_query_like}
\alias{build_query_like}
\title{Build a Like Query}
\usage{
build_query_like(
  fields = "*",
  distinct = FALSE,
  schema,
  table,
  where_like_field,
  where_like_value,
  case_insensitive = TRUE,
  limit_n = NULL
)
}
\description{
A \verb{Like Query} is one where a search for a value is performed with flanking wildcards.
}
\seealso{
\code{\link[SqlRender]{render}}
}
