% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build1.R
\name{build_query_like}
\alias{build_query_like}
\title{Writes a Like SQL Query}
\usage{
build_query_like(
  fields = "*",
  distinct = FALSE,
  schema,
  tableName,
  whereLikeField,
  whereLikeValue,
  caseInsensitive = TRUE,
  limit_n = NULL
)
}
\value{
a SQL Query as a character string.
}
\description{
Writes a Like SQL Query
}
