% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_query_regex}
\alias{build_query_regex}
\title{Build a Regex Query}
\usage{
build_query_regex(
  fields = "*",
  distinct = FALSE,
  schema,
  table,
  where_regex_field,
  regex,
  invert = FALSE,
  case_insensitive = TRUE,
  limit,
  random
)
}
\description{
Build a query that searches for a regular expression.
}
\seealso{
\code{\link[cli]{cli_alert}}
\code{\link[SqlRender]{render}}
}
