% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{buildJoinQuery}
\alias{buildJoinQuery}
\title{Build a JOIN SQL Query}
\usage{
buildJoinQuery(
  fields = "*",
  distinct = FALSE,
  schema,
  tableName,
  column,
  joinOnSchema,
  joinOnTableName,
  joinOnColumn,
  joinType = "LEFT",
  whereInField = NULL,
  whereInVector = NULL,
  whereNotInField = NULL,
  whereNotInVector = NULL,
  caseInsensitive = TRUE,
  n = NULL,
  n_type = c("limit", "random")
)
}
\value{
SQL statement as a character string.
}
\description{
A SQL query is built using the given arguments. Currently, only 1 whereIn and whereNot in parameters can be set.
}
