% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build1.R
\name{buildQuery}
\alias{buildQuery}
\title{Build a SQL Query}
\usage{
buildQuery(
  fields = "*",
  distinct = FALSE,
  schema,
  tableName,
  whereInField = NULL,
  whereInVector = NULL,
  whereNotInField = NULL,
  whereNotInVector = NULL,
  caseInsensitive = TRUE,
  n = NULL,
  n_type = c("limit", "random")
)
}
\value{
SQL statement as a character string.
}
\description{
A SQL query is built using the given arguments. Currently, only 1 whereIn and whereNot in parameters can be set.
}
