% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{buildQueryString}
\alias{buildQueryString}
\title{Writes a SQL Query Loop}
\usage{
buildQueryString(
  fields = "*",
  distinct = FALSE,
  schema,
  tableName,
  whereLikeField,
  string,
  split,
  caseInsensitive = TRUE,
  limit_n = NULL
)
}
\value{
a SQL Query as a character string.
}
\description{
This function writes a SQL Query that loops over the words in a string.
}
