% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn-env.R
\name{close_conn}
\alias{close_conn}
\title{Close Connections in \code{pg13_connection_env}}
\usage{
close_conn(..., all = FALSE, verbose = TRUE)
}
\arguments{
\item{...}{Names of the connections to close.}

\item{all}{Close all connections in the \code{pg13_connection_env} environment?}
}
\description{
Close the specified list of connection objects in the \code{pg13_connection_env} or close all connections, a feature is handy for making sure all connections are closed at the end of a session.
}
\seealso{
\code{\link[rlang]{list2}}
}
