% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{constructSchemaTableName}
\alias{constructSchemaTableName}
\title{Construct schemaTableName}
\usage{
constructSchemaTableName(schema, tableName)
}
\description{
(Deprecated)  construct schemaTableName from the schema and tableName
}
\keyword{internal}
