% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructWhereLike.R
\name{constructWhereLike}
\alias{constructWhereLike}
\title{Render "WHERE x IN y" Component}
\usage{
constructWhereLike(field, term)
}
\arguments{
\item{field}{Single field to be filtered for}

\item{vector}{vector of values that the SQL query is filtering for}
}
\description{
This is a non-terminal render, meaning that the SQL component will not be terminated with a semicolon in order to construct complex SQL queries.
}
