% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec.R
\name{constructWhereLowerLike}
\alias{constructWhereLowerLike}
\title{Render WHERE lowercase x LIKE y}
\usage{
constructWhereLowerLike(field, term)
}
\arguments{
\item{field}{Single field to be filtered for}

\item{term}{Character string that the field parameter is searched for.}
}
\description{
Render WHERE lowercase x LIKE y
}
