% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{create_log}
\alias{create_log}
\title{FUNCTION_TITLE}
\usage{
create_log(
  datetime_field,
  conn,
  schema,
  log_table,
  verbose = TRUE,
  render_sql = TRUE
)
}
\arguments{
\item{datetime_field}{PARAM_DESCRIPTION}

\item{conn}{PARAM_DESCRIPTION}

\item{schema}{PARAM_DESCRIPTION}

\item{log_table}{PARAM_DESCRIPTION}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}

\item{render_sql}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[SqlRender]{render}}
}
