% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_funs.R
\name{create_schema}
\alias{create_schema}
\title{Create a Schema}
\usage{
create_schema(
  conn,
  conn_fun,
  schema,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  ...
)
}
\description{
Create a Schema
}
