% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_schema}
\alias{create_schema}
\title{Create a Schema}
\usage{
create_schema(
  conn,
  conn_fun,
  schema,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  ...
)
}
\description{
Create a Schema
}
\examples{
library(pg13)
create_test_schema <-
        function(conn) {

                if (!schema_exists(conn = conn,
                                   schema = "test_schema")) {

                        cli::cli_rule("Create 'test_schema' Schema")

                        create_schema(conn = conn,
                                      schema = "test_schema")

                }
        }


conn <- local_connect(dbname = "pg13_test")
create_test_schema(conn = conn)
write_table(conn = conn,
            schema = "test_schema",
            table_name = "test_table",
            drop_existing = TRUE,
            data = data.frame(A = 1:3, B = letters[1:3]))

write_table(conn = conn,
            schema = "test_schema",
            table_name = "test_table2",
            drop_existing = TRUE,
            data = data.frame(C = rep(NA, 3), D = c(TRUE, FALSE, FALSE)))

write_table(conn = conn,
            schema = "test_schema",
            table_name = "test_table3",
            drop_existing = TRUE,
            data = data.frame(E = c(1.25, 343.31341, 5),
                              G = c(Sys.Date(), Sys.Date()-100, Sys.Date()-1000)))

drop_schema(conn = conn,
            schema = "test_schema",
            cascade = TRUE)

dc(conn = conn)
}
\seealso{
Other schema functions: 
\code{\link{drop_cascade}()},
\code{\link{drop_if_exists}()},
\code{\link{drop_schema}()}

Other create functions: 
\code{\link{create_table_from_df}()},
\code{\link{create_table}()},
\code{\link{draft_create_table_from_df}()},
\code{\link{draft_create_table}()}
}
\concept{create functions}
\concept{schema functions}
