% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table.R
\name{create_table}
\alias{create_table}
\title{Create table in the given database}
\usage{
create_table(dataframe, table_name, dbname, schema = NULL)
}
\arguments{
\item{dataframe}{data to write to table}

\item{table_name}{name of new table}

\item{dbname}{name of database for new table}

\item{schema}{name of schema for new table}
}
\description{
This function connects, writes a dataframe to a table using the DBI::dbWriteTable function, and disconnects.
}
