% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table.R
\name{create_table}
\alias{create_table}
\title{Create a Table}
\usage{
create_table(
  conn,
  schema,
  table_name,
  if_not_exists = TRUE,
  ...,
  verbose = TRUE,
  render_sql = TRUE
)
}
\arguments{
\item{...}{Named vector of field names and their corresponding data definition.}
}
\description{
Create a Table
}
\seealso{
\code{\link[rlang]{list2}}
\code{\link[SqlRender]{render}}
\code{\link[pg13]{send}}
}
