% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTable.R, R/create_table.R
\name{draft_create_table}
\alias{draft_create_table}
\alias{create_table}
\title{Draft SQL to Create a Table}
\usage{
draft_create_table(schema, tableName, if_not_exists = TRUE, ...)

create_table(
  conn,
  schema,
  tableName,
  if_not_exists = TRUE,
  ...,
  verbose = TRUE,
  render_sql = TRUE
)

draft_create_table(schema, tableName, if_not_exists = TRUE, ...)

create_table(
  conn,
  schema,
  tableName,
  if_not_exists = TRUE,
  ...,
  verbose = TRUE,
  render_sql = TRUE
)
}
\arguments{
\item{...}{Named vector of field names and their corresponding data definition.}
}
\description{
Draft a SQL that creates a table using DDL derived from a dataframe. Drafting is stopped if the given tablename or fields in the dataframe are a SQL reserve word.

Draft a SQL that creates a table using DDL derived from a dataframe. Drafting is stopped if the given tablename or fields in the dataframe are a SQL reserve word.
}
\seealso{
\code{\link[rlang]{list2}}
\code{\link[SqlRender]{render}}

\code{\link[rlang]{list2}}
\code{\link[SqlRender]{render}}
\code{\link[pg13]{send}}

\code{\link[rlang]{list2}}
\code{\link[SqlRender]{render}}

\code{\link[rlang]{list2}}
\code{\link[SqlRender]{render}}
\code{\link[pg13]{send}}
}
