% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_table}
\alias{create_table}
\title{Create a Table}
\usage{
create_table(
  conn,
  conn_fun = "pg13::local_connect()",
  schema,
  table_name,
  if_not_exists = TRUE,
  ...,
  verbose = TRUE,
  render_sql = TRUE
)
}
\arguments{
\item{...}{Named vector of field names and their corresponding data definition.}
}
\description{
Create a Table
}
\examples{
library(pg13)
create_test_schema <-
        function(conn) {

                if (!schema_exists(conn = conn,
                                   schema = "test_schema")) {

                        create_schema(conn = conn,
                                      schema = "test_schema")

                }
        }

conn <- local_connect(dbname = "pg13_test")
create_test_schema(conn = conn)

# Write a table
write_table(conn = conn,
            schema = "test_schema",
            table_name = "test_table",
            drop_existing = TRUE,
            data = data.frame(A = 1:3, B = letters[1:3]))

# Write the same table using create_table() instead
create_table(conn = conn,
             schema = "test_schema",
             table_name = "test_table_b",
             if_not_exists = TRUE,
             A = "integer",
             B = "varchar(1)")

append_table(conn = conn,
             schema = "test_schema",
             table = "test_table_b",
             data = data.frame(A = 1:3, B = letters[1:3]))

# Under the hood is the draft_create_table()
draft_create_table(schema = "test_schema",
                   table_name = "test_table_b",
                   A = "integer",
                   B = "varchar(1)")

# The DDL can be automatically discerned using create_table_from_df()
create_table_from_df(conn = conn,
                     schema = "test_schema",
                     table_name = "test_table_c",
                     data = data.frame(A = 1:3, B = letters[1:3]))

# Under the hood is the draft_create_table_from_df()
draft_create_table_from_df(schema = "test_schema",
                           table_name = "test_table_c",
                           data = data.frame(A = 1:3, B = letters[1:3]))

drop_schema(conn = conn,
            schema = "test_schema",
            cascade = TRUE)

dc(conn = conn)
}
\seealso{
\code{\link[rlang]{list2}}
\code{\link[SqlRender]{render}}
\code{\link[pg13]{send}}

Other create functions: 
\code{\link{create_schema}()},
\code{\link{create_table_from_df}()},
\code{\link{draft_create_table_from_df}()},
\code{\link{draft_create_table}()}

Other table functions: 
\code{\link{appendTable}()},
\code{\link{append_table}()},
\code{\link{create_table_from_df}()},
\code{\link{drop_all_staging_tables}()},
\code{\link{drop_table_batch}()},
\code{\link{drop_table}()},
\code{\link{read_table}()},
\code{\link{searchTable}()},
\code{\link{search_table}()},
\code{\link{write_staging_table}()},
\code{\link{write_table}()}
}
\concept{create functions}
\concept{table functions}
