% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{createDB}
\alias{createDB}
\title{Create a New Database}
\usage{
createDB(conn, dbname, ...)
}
\arguments{
\item{conn}{Connection object}

\item{dbname}{Database name.}

\item{...}{Additional arguments passed to \code{DatabaseConnector::dbSendStatement()}.}
}
\description{
Create a New Database
}
\keyword{internal}
