% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{dc}
\alias{dc}
\title{Disconnect a Connection}
\usage{
dc(conn, ..., verbose = TRUE, remove = FALSE)
}
\arguments{
\item{conn}{Connection object}

\item{...}{Additional arguments passed to \code{\link[DatabaseConnector]{dbDisconnect}}.}

\item{remove}{If TRUE, the Connection object argument is removed from the Parent Environment.}
}
\description{
Disconnect a Postgres Connection object with the option of removing the object from the parent environment.
}
