% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{dc}
\alias{dc}
\title{Disconnect a Connection}
\usage{
dc(conn, verbose = TRUE, remove = FALSE)
}
\arguments{
\item{verbose}{If TRUE, returns a console message when the connection has been closed.}

\item{remove}{If TRUE, the Connection object argument is removed from the parent environment.}

\item{...}{Additional arguments passed to \code{\link[DatabaseConnector]{dbDisconnect}}.}
}
\description{
Disconnect a Postgres Connection object with the option of removing the object from the parent environment.
}
\seealso{
Other connection functions: 
\code{\link{conn_db}()},
\code{\link{connect_ff}()},
\code{\link{connect}()},
\code{\link{dc0}()},
\code{\link{quietly.conn_db}()}
}
\concept{connection functions}
