% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R, R/deprecated.R
\name{dc}
\alias{dc}
\title{Disconnect a Connection}
\usage{
dc(conn, ..., verbose = TRUE, remove = FALSE)

dc(conn, ..., verbose = TRUE, remove = FALSE)
}
\arguments{
\item{conn}{Connection object}

\item{...}{Additional arguments passed to \code{\link[DatabaseConnector]{dbDisconnect}}.}

\item{verbose}{If TRUE, returns a console message when the connection has been closed.}

\item{remove}{If TRUE, the Connection object argument is removed from the parent environment.}
}
\description{
Disconnect a Postgres Connection object with the option of removing the object from the parent environment.

Disconnect a Postgres Connection object with the option of removing the object from the parent environment.
}
\keyword{internal}
