% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{draft_base}
\alias{draft_base}
\title{Construct a base SQL query}
\usage{
draft_base(fields = "*", distinct = FALSE, schema, table_name)
}
\arguments{
\item{distinct}{if TRUE, the DISTINCT fields will be selected for.}
}
\description{
Construct the SELECT {fields} FROM {schema}.{table} base query. The SQL Statement is not terminated for the addition of other parameters generated using "construct" functions.
}
