% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{draft_create_table}
\alias{draft_create_table}
\title{Draft SQL to Create a Table}
\usage{
draft_create_table(schema, table_name, if_not_exists = TRUE, ...)
}
\arguments{
\item{...}{Named vector of field names and their corresponding data definition.}
}
\description{
Draft a SQL that creates a table using DDL derived from a dataframe. Drafting is stopped if the given tablename or fields in the dataframe are a SQL reserve word.
}
\seealso{
\code{\link[rlang]{list2}}
\code{\link[SqlRender]{render}}

Other create functions: 
\code{\link{create_table_from_df}()},
\code{\link{create_table}()},
\code{\link{draft_create_table_from_df}()}

Other draft functions: 
\code{\link{create_table_from_df}()},
\code{\link{draft_create_table_from_df}()}
}
\concept{create functions}
\concept{draft functions}
