% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table.R
\name{draft_create_table_from_df}
\alias{draft_create_table_from_df}
\title{Draft the SQL to Create Table}
\usage{
draft_create_table_from_df(schema, table_name, data, if_not_exists = TRUE)
}
\description{
Draft a SQL that creates a table using DDL derived from a dataframe. Drafting is stopped if the given tablename or fields in the dataframe are a SQL reserve word.
}
\seealso{
\code{\link[forcats]{fct_collapse}}
}
