% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decamel-draft.R
\name{draft_in}
\alias{draft_in}
\title{Render y in "WHERE x IN (y)"}
\usage{
draft_in(vector)
}
\arguments{
\item{vector}{vector of values that the SQL query is filtering for}
}
\description{
This is a non-terminal render, meaning that the SQL component will not be terminated with a semicolon in order to construct complex SQL queries.
}
