% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{draft_join}
\alias{draft_join}
\title{Construct a join part of a SQL Statement}
\usage{
draft_join(
  schema,
  tableName,
  column,
  joinType = "LEFT",
  joinOnSchema,
  joinOnTableName,
  joinOnColumn
)
}
\description{
Construct {join type} JOIN {schema.table.column2} ON {schema.table.column2} = {schema.table.column1} part of the sql statement
}
\seealso{
Other draft functions: 
\code{\link{draft_base}()},
\code{\link{draft_create_table_from_df}()},
\code{\link{draft_create_table}()},
\code{\link{draft_in}()},
\code{\link{draft_limit}()},
\code{\link{draft_random}()},
\code{\link{draft_table_path}()},
\code{\link{draft_where_in}()},
\code{\link{draft_where_like}()},
\code{\link{draft_where_lower_in}()},
\code{\link{draft_where_lower_like}()},
\code{\link{draft_where_lower_not_in}()},
\code{\link{draft_where_not_in}()},
\code{\link{paste_wheres}()},
\code{\link{table.path}()}
}
\concept{draft functions}
