% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{draft_limit}
\alias{draft_limit}
\title{Construct LIMIT}
\usage{
draft_limit(n)
}
\arguments{
\item{n}{rows to limit to}
}
\description{
This is a non-terminal render, meaning that the SQL component will not be terminated with a semicolon in order to construct complex SQL queries.
}
