% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_funs.R
\name{drop_cascade}
\alias{drop_cascade}
\title{Drop a Postgres schema}
\usage{
drop_cascade(conn, schema, ...)

drop_cascade(conn, schema, ...)
}
\arguments{
\item{...}{Additional arguments passed to the DatabaseConnector::dbSendStatement function}
}
\description{
Drop a schema if it exists.

Drop a schema if it exists.
}
