% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/schema_funs.R
\name{drop_schema}
\alias{drop_schema}
\title{Drop a Postgres schema}
\usage{
drop_schema(
  conn,
  conn_fun,
  schema,
  cascade = FALSE,
  if_exists = TRUE,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  ...
)

drop_schema(
  conn,
  conn_fun,
  schema,
  cascade = FALSE,
  if_exists = TRUE,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  ...
)
}
\arguments{
\item{...}{Additional arguments passed to the DatabaseConnector::dbSendStatement function}
}
\description{
Drop a schema if it exists.

Drop a schema if it exists.
}
\keyword{internal}
