% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decamel-deprec.R, R/decamel-schema_funs.R
\name{drop_schema}
\alias{drop_schema}
\title{Drop a Postgres schema}
\usage{
drop_schema(conn, schema, cascade = FALSE, if_exists = TRUE, ...)

drop_schema(conn, schema, cascade = FALSE, if_exists = TRUE, ...)
}
\arguments{
\item{...}{Additional arguments passed to the DatabaseConnector::dbSendStatement function}
}
\description{
Drop a schema if it exists.

Drop a schema if it exists.
}
