% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_table.R
\name{drop_table}
\alias{drop_table}
\title{Drop table in the given database}
\usage{
drop_table(table_name, dbname, schema = NULL)
}
\arguments{
\item{table_name}{name of new table}

\item{dbname}{name of database for new table}

\item{schema}{name of schema for new table}
}
\description{
Drop table in the given database
}
