% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropSchema.R
\name{dropSchema}
\alias{dropSchema}
\title{Drop a table in a Postgres schema}
\usage{
dropSchema(conn, schema, cascade = FALSE, tableName, ...)
}
\arguments{
\item{...}{Additional arguments passed to the DatabaseConnector::dbSendStatement function}
}
\description{
Drop a table if it exists.
}
