% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{dropSchema}
\alias{dropSchema}
\title{Drop a Postgres schema}
\usage{
dropSchema(conn, schema, cascade = FALSE, if_exists = TRUE, ...)

dropSchema(conn, schema, cascade = FALSE, if_exists = TRUE, ...)
}
\arguments{
\item{...}{Additional arguments passed to the DatabaseConnector::dbSendStatement function}
}
\description{
Drop a schema if it exists.

Drop a schema if it exists.
}
\keyword{internal}
