% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{dropTable}
\alias{dropTable}
\title{Drop a Table}
\usage{
dropTable(
  conn,
  conn_fun,
  schema,
  tableName,
  if_exists = TRUE,
  verbose = TRUE,
  render_sql = TRUE,
  ...
)
}
\arguments{
\item{if_exists}{If TRUE, the table will be dropped only if it exists.}

\item{...}{Additional arguments passed to \code{DatabaseConnector::dbSendStatement()}}
}
\description{
Drop a Table
}
\keyword{internal}
