% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{dropTable}
\alias{dropTable}
\title{Drop a table in a Postgres schema}
\usage{
dropTable(conn, schema, tableName, if_exists = TRUE, ...)
}
\arguments{
\item{...}{Additional arguments passed to the DatabaseConnector::dbSendStatement function}
}
\description{
Drop a table if it exists.
}
