% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join2}
\alias{join2}
\title{Second Generation Join}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{column}{PARAM_DESCRIPTION, Default: NULL}

\item{omopSchema}{PARAM_DESCRIPTION}

\item{omopTable}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}

\item{writeSchema}{PARAM_DESCRIPTION}

\item{join_type}{PARAM_DESCRIPTION, Default: 'LEFT'}

\item{case_insensitive}{PARAM_DESCRIPTION, Default: TRUE}

\item{render_sql}{PARAM_DESCRIPTION, Default: TRUE}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
The second generation join is a join where the added feature of joining on multiple columns with "AND" logic is implemented. Another change is that all the join columns can optionally be cast to varchar to simplify the sql render. See \code{\link{join}} for the first generation join.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[SqlRender]{render}}
}
