% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_stop_ff}
\alias{log_stop_ff}
\title{FUNCTION_TITLE}
\usage{
log_stop_ff(datetime_field, schema, log_table)
}
\arguments{
\item{datetime_field}{PARAM_DESCRIPTION}

\item{schema}{PARAM_DESCRIPTION}

\item{log_table}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{tibble}}
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}},\code{\link[dplyr]{reexports}}
}
