% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{ls_fields}
\alias{ls_fields}
\title{List Fields}
\usage{
ls_fields(conn, schema, tableName, verbose = TRUE, render_sql = TRUE)
}
\description{
Fields for the given table are returned in lowercase.
}
\seealso{
Other list functions: 
\code{\link{ls_schema}()},
\code{\link{ls_tables}()}
}
\concept{list functions}
