% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{ls_schema}
\alias{ls_schema}
\title{List Schemas}
\usage{
ls_schema(conn, verbose = TRUE, render_sql = TRUE, render_only = FALSE)
}
\description{
List all the schemas in a database in lowercase.
}
\seealso{
Other list functions: 
\code{\link{lsFields}()},
\code{\link{lsSchema}()},
\code{\link{lsTables}()},
\code{\link{ls_db}()},
\code{\link{ls_fields}()},
\code{\link{ls_tables}()},
\code{\link{ls_users}()}
}
\concept{list functions}
