% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{ls_tables}
\alias{ls_tables}
\title{List Tables}
\usage{
ls_tables(conn, conn_fun, schema, verbose = TRUE, render_sql = TRUE)
}
\description{
List Tables
}
\seealso{
Other list functions: 
\code{\link{lsFields}()},
\code{\link{lsSchema}()},
\code{\link{lsTables}()},
\code{\link{ls_db}()},
\code{\link{ls_fields}()},
\code{\link{ls_schema}()}
}
\concept{list functions}
