% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{lsFields}
\alias{lsFields}
\title{List Fields}
\usage{
lsFields(conn, schema, tableName, verbose = TRUE, render_sql = TRUE)
}
\description{
Fields for the given table are returned in lowercase.
}
\seealso{
Other list functions: 
\code{\link{lsSchema}()},
\code{\link{lsTables}()},
\code{\link{ls_fields}()},
\code{\link{ls_schema}()},
\code{\link{ls_tables}()}
}
\concept{list functions}
\keyword{internal}
