% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{lsSchema}
\alias{lsSchema}
\title{List Schemas}
\usage{
lsSchema(conn, verbose = TRUE, render_sql = TRUE)
}
\description{
List all the schemas in a database in lowercase.
}
\seealso{
Other list functions: 
\code{\link{lsFields}()},
\code{\link{lsTables}()},
\code{\link{ls_db}()},
\code{\link{ls_fields}()},
\code{\link{ls_schema}()},
\code{\link{ls_tables}()}
}
\concept{list functions}
