% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_db.R
\name{makeLocalConnDetails}
\alias{makeLocalConnDetails}
\title{Make a local Postgres Connection Details object}
\usage{
makeLocalConnDetails(dbname = "athena", port = 5432)
}
\arguments{
\item{dbname}{Name of a local Postgres database, Default: 'athena'}

\item{port}{The port on the server to connect to, Default: 5432}
}
\value{
A connectionDetails class object that lists all the details needed to connect to a database.
}
\description{
Creating a connection to a Postgres Database involves first making a connectionDetails object that is a list of the credentials and other parameters to the target database followed by making the connection itself that is stored as connection object.
}
\details{
DETAILS
}
\section{Local Connection Details}{

This function returns a connectionDetails class object for functions that make the database connection within its caller environment, such as those belonging to the OHDSI Feature Extraction package. A connection can be made by passing the connectionDetails class object to the \code{\link[DatabaseConnector]{connect}}. This process can be simplified into a single function call with \code{\link{localConnect}}, which performs both operations but will not return a connectionDetails object if it is desired as a return value.
}

\section{Local Connection}{

This function combines the process of creating a connectionDetails object and making the connection in a single function call. If a connectionDetails object is required, see \code{\link{makeLocalConnDetails}}.
}

\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[DatabaseConnector]{createConnectionDetails}}
}
\concept{local_db}
