% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{makeLocalConnDetails}
\alias{makeLocalConnDetails}
\title{Make a local Postgres Connection Details object}
\usage{
makeLocalConnDetails(dbname = "athena", port = 5432)
}
\arguments{
\item{dbname}{Name of a local Postgres database, Default: 'athena'}

\item{port}{The port on the server to connect to, Default: 5432}
}
\value{
A connectionDetails class object that lists all the details needed to connect to a database.
}
\description{
Connect and disconnect to a local Postgres database that does not require credentialing. These functions automatically set the dbms and server arguments to "postgresql" and "localhost/{dbname}", leaving the database name (\code{dbname}) and port (\code{port}) as the only connection details that are modifiable by the user. Note that these functions cannot be used to connect to a local instance that requires a username and password.

Connect and disconnect to a local Postgres database that does not require credentialing. These functions automatically set the dbms and server arguments to "postgresql" and "localhost/{dbname}", leaving the database name (\code{dbname}) and port (\code{port}) as the only connection details that are modifiable by the user. Note that these functions cannot be used to connect to a local instance that requires a username and password.
}
\section{Local ConnectionDetails}{

This function returns a connectionDetails class object for functions that make the database connection within its caller environment, such as those belonging to the OHDSI Feature Extraction package. A connection can be made by passing the connectionDetails class object to the \code{\link[DatabaseConnector]{connect}}. This process can be simplified into a single function call with \code{\link{localConnect}}, which performs both operations but will not return a connectionDetails object if it is desired as a return value.


This function returns a connectionDetails class object for functions that make the database connection within its caller environment, such as those belonging to the OHDSI Feature Extraction package. A connection can be made by passing the connectionDetails class object to the \code{\link[DatabaseConnector]{connect}}. This process can be simplified into a single function call with \code{\link{localConnect}}, which performs both operations but will not return a connectionDetails object if it is desired as a return value.
}

\seealso{
\code{\link[DatabaseConnector]{createConnectionDetails}}

Other local functions: 
\code{\link{localConnect}()},
\code{\link{local_connect}()},
\code{\link{local}()},
\code{\link{make_local_conn_details}()}
}
\concept{local functions}
