% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn-env.R
\name{open_conn}
\alias{open_conn}
\title{Open a Connection in the \code{pg13_connection_env}}
\usage{
open_conn(conn_fun, conn_name, verbose = TRUE)
}
\arguments{
\item{conn_fun}{Function call for the connection as a character string such as "local_connect()".}

\item{conn_name}{Name of the connection object in the \code{pg13_connection_env} environment}
}
\description{
Connect to a database and assign it to the \code{pg13_connection_env}. This is a method to keep mulitple database connections organized during a crosswalk project. Existing connection objects will not be overwritten If the connection object of the provided name already exists. If the \code{pg13_connection_env} does not exist in the Global Environment, it will also be created at this point beforehand.
}
\seealso{
\code{\link[secretary]{c("typewrite", "typewrite")}},\code{\link[secretary]{italicize}}
\code{\link[rlang]{parse_expr}}

Other pg13 environment functions: 
\code{\link{[[<-,pg_env,character,missing,ANY-method}},
\code{\link{close_conn}()},
\code{\link{pg_env-class}},
\code{\link{see_conn}()}

Other pg13 class connection functions: 
\code{\link{close_conn}()},
\code{\link{see_conn}()}
}
\concept{pg13 class connection functions}
\concept{pg13 environment functions}
