% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn-env.R
\name{[[<-,pg_env,character,missing,ANY-method}
\alias{[[<-,pg_env,character,missing,ANY-method}
\title{Methods: PG13 environment}
\usage{
\S4method{[[}{pg_env,character,missing,ANY}(x, i) <- value
}
\description{
Methods: PG13 environment
}
\seealso{
Other pg13 environment functions: 
\code{\link{close_conn}()},
\code{\link{open_conn}()},
\code{\link{pg_env-class}},
\code{\link{see_conn}()}

Other pg13 class functions: 
\code{\link{pg_env-class}}
}
\concept{pg13 class functions}
\concept{pg13 environment functions}
