% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/utils-general.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs

lhs \%>\% rhs
}
\description{
See \code{magrittr::\link[magrittr:pipe]{\%>\%}} for details.

See \code{magrittr::\link[magrittr:pipe]{\%>\%}} for details.
}
\keyword{internal}
