% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{query}
\alias{query}
\title{Query}
\usage{
query(
  conn,
  conn_fun = "pg13::local_connect()",
  checks = c("conn_status", "conn_type", "rows", "total_rows"),
  sql_statement,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  log_file = "",
  append_log = TRUE,
  sep_log = "\\n",
  sql_style = c("inline", "chunk"),
  rmd_file = "",
  sql_file = "",
  warn_no_rows = "deprecated",
  warnMissingParameters = TRUE,
  ...
)
}
\arguments{
\item{conn}{Connection object}

\item{sql_statement}{A SQL Statement.}

\item{verbose}{If TRUE, details on the activity are returned in the console, such as when the querying starting and was completed.}

\item{render_sql}{If TRUE, the SQL statement for the operation is returned in the console.}

\item{warn_no_rows}{If TRUE, a warning is given that query has returned 0 rows.}
}
\description{
Query
}
\seealso{
\code{\link[rlang]{parse_expr}}
}
