% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{query}
\alias{query}
\title{Query}
\usage{
query(
  conn,
  conn_fun,
  sql_statement,
  verbose = TRUE,
  render_sql = TRUE,
  warn_no_rows = "deprecated",
  render_only = FALSE,
  ...
)
}
\description{
Query
}
\seealso{
\code{\link[rlang]{parse_expr}}
}
