% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Send Query to any Postgres connection}
\usage{
query(conn, sql_statement, render_sql = TRUE, ...)
}
\arguments{
\item{conn}{Connection object}

\item{sql_statement}{SQL Query}

\item{...}{Additional arguments to pass onto \code{\link[DatabaseConnector]{dbGetQuery}}}
}
\description{
Send Query to any Postgres connection
}
