% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{rm_if_closed}
\alias{rm_if_closed}
\title{Remove a Closed Connection Object}
\usage{
rm_if_closed(conn)
}
\arguments{
\item{conn}{Postgres connection object}
}
\description{
(Deprecated) This function removes a connection object from the Global Environment if it is closed.
}
