% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decamel-connection.R
\name{rm_if_closed}
\alias{rm_if_closed}
\title{Remove a Closed Connection Object}
\usage{
rm_if_closed(conn)
}
\arguments{
\item{conn}{A connection object returned by \code{connect()}.}
}
\description{
If a connection is not open, it is removed from the parent environment from which this function is called.
}
