% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isClosed.R
\name{rmIfClosed}
\alias{rmIfClosed}
\title{Remove a Closed Connection Object}
\usage{
rmIfClosed(conn)
}
\arguments{
\item{conn}{Postgres connection object}
}
\description{
This function removes a onnection object from the Global Environment if it is closed.
}
