% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{searchTable}
\alias{searchTable}
\title{Search a Table for a Value}
\usage{
searchTable(conn, schema, tableName, values, case_insensitive = TRUE, ...)
}
\description{
If the Field to query is unknown, this function can be used to search and query all Fields or a subset of Fields for a given value. The \code{value} is converted to character and the Field is cast to varchar in the query and as a result, the query times may be long which may require providing a subset of Fields as an argument rather than searching the entire Table in many cases.
}
