% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{searchTable}
\alias{searchTable}
\title{Search a Table for a Value}
\usage{
searchTable(
  conn,
  schema,
  tableName,
  ...,
  values,
  case_insensitive = TRUE,
  verbose = TRUE,
  render_sql = TRUE
)
}
\arguments{
\item{...}{(Optional) Character strings of 1 or more fields in the table to search in.}

\item{values}{Vector of length 1 or greater to search for.}

\item{case_insensitive}{If TRUE, both sides of the query are converted to lowercase.}
}
\description{
Loop a query for a set of one or more values in a table across all the existing fields or optionally, a subset of the fields. Both the values and the table fields are ensured compatibility by 1. Converting each value in the \code{values} argument to the character class and 2. Casting each table field as varchar in the query.
}
\examples{
conn <- localConnect(dbname = "athena",
                     port = 5432)

searchTable(conn = conn,
            schema = "public",
            tableName = "concept",
            values = "Aspirin",
            case_insensitive = TRUE)


searchTable(conn = conn,
            schema = "public",
            tableName = "concept",
            values = "Aspirin",
            case_insensitive = FALSE)


searchTable(conn = conn,
            schema = "public",
            tableName = "concept",
            values = "Aspirin",
            case_insensitive = FALSE,
            verbose = FALSE)

searchTable(conn = conn,
            schema = "public",
            tableName = "concept",
            values = "Aspirin",
            case_insensitive = FALSE,
            verbose = FALSE,
            render_sql = TRUE,
            "concept_id",
            "concept_name")


searchTable(conn = conn,
            schema = "public",
            tableName = "concept",
            c("concept_id", "concept_name"),
            values = "Aspirin",
            case_insensitive = FALSE,
            verbose = FALSE,
            render_sql = TRUE)

dc(conn = conn,
   remove = TRUE)
}
\seealso{
Other table functions: 
\code{\link{appendTable}()},
\code{\link{append_table}()},
\code{\link{search_table}()}
}
\concept{table functions}
\keyword{internal}
