% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn-env.R
\name{see_conn}
\alias{see_conn}
\title{See the Connections in \code{pg13_connection_env}}
\usage{
see_conn()
}
\description{
See the connection objects in the \code{pg13_connection_env} along with their open status.
}
\seealso{
\code{\link[rlang]{list2}}
}
