% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{send}
\alias{send}
\title{Send a SQL Statement to Postgres}
\usage{
send(
  conn,
  conn_fun,
  sql_statement,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  ...
)
}
\arguments{
\item{conn}{Connection object}

\item{sql_statement}{SQL to send}

\item{...}{Additional arguments to pass onto the DatabaseConnector::dbSendStatement function}
}
\description{
Send a SQL Statement to Postgres
}
