% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{send}
\alias{send}
\title{Send a SQL Statement to Postgres}
\usage{
send(
  conn,
  conn_fun = "pg13::local_connect()",
  sql_statement,
  log_file = "",
  append_log = TRUE,
  sep_log = "\\n",
  checks = c("conn_status", "conn_type"),
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  sql_style = c("inline", "chunk"),
  rmd_file = "",
  sql_file = "",
  warnOnMissingParameters = TRUE,
  ...
)
}
\description{
Send a SQL Statement to Postgres
}
