% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_funs.R
\name{sendList}
\alias{sendList}
\title{Send a List of SQL Statements}
\usage{
sendList(
  conn,
  sqlList,
  verbose = TRUE,
  progressBar = TRUE,
  skipErrors = TRUE,
  errorFile = NULL,
  ...
)
}
\arguments{
\item{conn}{Connection object}

\item{sqlList}{list object of queries}

\item{verbose}{If TRUE, the SQL is printed back before executing}

\item{progressBar}{If TRUE, a progress bar is returned in the console.}

\item{skipErrors}{If TRUE, if a SQL execution returns an error, the statement is printed back in red in the console and the iteration will proceed to the next sql statement in line.}

\item{errorFile}{(optional) path to the error file where any errors are written to if not null.}

\item{...}{Additional arguments to pass onto the DatabaseConnector::dbGetQuery function}
}
\description{
Iteratively send a list of SQL Statements such as the object returned by \code{\link{parseSQL}}.
}
