% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/draft.R
\name{table.path}
\alias{table.path}
\title{Construct schemaTableName}
\usage{
table.path(schema, table_name)

table.path(schema, table_name)
}
\description{
construct schemaTableName from the schema and tableName

construct schemaTableName from the schema and tableName
}
\seealso{
Other draft functions: 
\code{\link{draft_base}()},
\code{\link{draft_create_table_from_df}()},
\code{\link{draft_create_table}()},
\code{\link{draft_in}()},
\code{\link{draft_join}()},
\code{\link{draft_limit}()},
\code{\link{draft_random}()},
\code{\link{draft_table_path}()},
\code{\link{draft_where_in}()},
\code{\link{draft_where_like}()},
\code{\link{draft_where_lower_in}()},
\code{\link{draft_where_lower_like}()},
\code{\link{draft_where_lower_not_in}()},
\code{\link{draft_where_not_in}()},
\code{\link{paste_wheres}()}
}
\concept{draft functions}
