% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decamel-list.R, R/list.R
\name{table_exists}
\alias{table_exists}
\title{Does a table exist?}
\usage{
table_exists(conn, schema, tableName)

table_exists(conn, schema, tableName)
}
\arguments{
\item{conn}{A connection object returned by \code{connect()}.}

\item{schema}{The target schema for the operation.}

\item{tableName}{The target table in the \code{schema} for an operation or for writing functions, the name of the table that will be created.}
}
\description{
Does a table exist?

Does a table exist?
}
\seealso{
Other logical functions: 
\code{\link{field_exists}()},
\code{\link{schema_exists}()}

Other logical functions: 
\code{\link{field_exists}()},
\code{\link{schema_exists}()}
}
\concept{logical functions}
