% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exists.R
\name{table_exists}
\alias{table_exists}
\title{Does a table exist?}
\usage{
table_exists(conn, schema, table_name)
}
\arguments{
\item{conn}{Connection object}

\item{schema}{The target schema for the operation.}
}
\description{
Does a table exist?
}
\seealso{
Other logical functions: 
\code{\link{db_exists}()},
\code{\link{field_exists}()},
\code{\link{schema_exists}()}
}
\concept{logical functions}
