% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/utils-typewrite.R
\name{typewrite_activity}
\alias{typewrite_activity}
\title{Typewrite Activity}
\usage{
typewrite_activity(activity)

typewrite_activity(activity)
}
\arguments{
\item{activity}{Character string.}
}
\description{
Return the provided text in yellow along with a
timestamp.
}
\seealso{
\code{\link[secretary]{c("typewrite", "yellowTxt")}}
}
