% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decamel-table.R
\name{write_staging_table}
\alias{write_staging_table}
\title{Write a Table that is Dropped On Exit}
\usage{
write_staging_table(
  conn,
  conn_fun,
  schema,
  tableName,
  data,
  drop_existing = FALSE,
  verbose = TRUE,
  render_sql = TRUE,
  ...
)
}
\description{
A Staging Table is one that is dropped on exit in the parent frame from which the function is being called.
}
