% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{write_table}
\alias{write_table}
\title{Write a Table}
\usage{
write_table(
  conn,
  conn_fun,
  schema,
  tableName,
  data,
  drop_existing = FALSE,
  verbose = TRUE,
  render_sql = TRUE,
  ...
)
}
\arguments{
\item{...}{Additional arguments passed to \code{DatabaseConnector::dbWriteTable()}}
}
\description{
Unlike the dropTable and renameTable functions, this function is a wrapper around the \code{DatabaseConnector::dbWriteTable()} function rather than one where a SQL statement is rendered using the SqlRender package. This function that converts all inputs to a dataframe, especially in cases where the input is a tibble, in which case an error would be thrown when writing.
}
