% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{writeTable}
\alias{writeTable}
\title{Write a Table}
\usage{
writeTable(
  conn,
  conn_fun,
  schema,
  tableName,
  data,
  drop_existing = FALSE,
  verbose = TRUE,
  render_sql = TRUE,
  ...
)
}
\arguments{
\item{conn}{Connection object}

\item{schema}{The target schema for the operation.}

\item{tableName}{The target table in the \code{schema} for an operation or for writing functions, the name of the table that will be created.}

\item{data}{A dataframe or tibble.}

\item{verbose}{If TRUE, details on the activity are returned in the console, such as when the querying starting and was completed.}

\item{render_sql}{If TRUE, the SQL statement for the operation is returned in the console.}

\item{...}{Additional arguments passed to \code{DatabaseConnector::dbWriteTable()}}
}
\description{
Unlike the dropTable and renameTable functions, this function is a wrapper around the \code{DatabaseConnector::dbWriteTable()} function rather than one where a SQL statement is rendered using the SqlRender package. This function that converts all inputs to a dataframe, especially in cases where the input is a tibble, in which case an error would be thrown when writing.
}
